% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{add_attachment_xlsx}
\alias{add_attachment_xlsx}
\title{Add attachment to email}
\usage{
add_attachment_xlsx(email, session_id = NULL)
}
\arguments{
\item{email}{The email message object, as created by the \code{\link[blastula:compose_email]{compose_email()}}
function. The object's class is \code{email_message}.}

\item{session_id}{Character vector to identify current shiny session}
}
\description{
Add attachment to email
}
