% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{add_bib_match}
\alias{add_bib_match}
\title{Match table from bib table}
\usage{
add_bib_match(paper, min_score = 50)
}
\arguments{
\item{paper}{a paper or paperlist object}

\item{min_score}{minimal score that is taken to be a reliable match}
}
\value{
the paper or paperlist with bib_match table added
}
\description{
Match table from bib table
}
\examples{
\dontrun{
paper <- demopaper()
paper$bib_match <- NULL # remove existing
paper2 <- add_bib_match(paper)
paper2$bib_match
}
}
