% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{apa_authors}
\alias{apa_authors}
\title{Get author info from JATS APA-DTD type XML}
\usage{
apa_authors(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
an author list
}
\description{
Get author info from JATS APA-DTD type XML
}
\keyword{internal}
