% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{apa_info}
\alias{apa_info}
\title{Get article info from JATS APA-DTD type XML}
\usage{
apa_info(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
an info list
}
\description{
Get article info from JATS APA-DTD type XML
}
\keyword{internal}
