% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-aspredicted.R
\name{aspredicted_info}
\alias{aspredicted_info}
\title{Retrieve info from AsPredicted by URL}
\usage{
aspredicted_info(ap_url, id_col = 1, wait = 1)
}
\arguments{
\item{ap_url}{an AsPredicted URL, or a table containing them (e.g., as created by \code{aspredicted_links()})}

\item{id_col}{the index or name of the column that contains AsPredicted URLs, if id is a table}

\item{wait}{wait time in seconds}
}
\value{
a data frame of information
}
\description{
Retrieve info from AsPredicted by URL
}
