% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-aspredicted.R
\name{aspredicted_links}
\alias{aspredicted_links}
\title{Find AsPredicted Links in Papers}
\usage{
aspredicted_links(paper)
}
\arguments{
\item{paper}{a paper object or paperlist object}
}
\value{
a table with the AsPredicted url in the first (href) column
}
\description{
Find AsPredicted Links in Papers
}
\examples{
\dontrun{
psychsci <- papers_load("psychsci", cache = TRUE)
aspredicted_links(psychsci)
}
}
