% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{auth_cr}
\alias{auth_cr}
\title{Authorise crossref requests}
\usage{
auth_cr()
}
\description{
Tries \code{\link[crlite:polite]{crlite::get_cr_mailto()}} and \code{\link[crlite:plus]{crlite::get_cr_token()}} or sets \code{character(1)} otherwise.
}
\seealso{
Other helpers: 
\code{\link{accumulate_pred_trans}()},
\code{\link{auth_mailjet}()}
}
\concept{helpers}
