% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{auth_mailjet}
\alias{auth_mailjet}
\title{Authorise mailjet requests}
\usage{
auth_mailjet()
}
\description{
Tries the env var, then keyring and sets as env var.
Otherwise throws warning.
}
\seealso{
Other helpers: 
\code{\link{accumulate_pred_trans}()},
\code{\link{auth_cr}()}
}
\concept{helpers}
