% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{bib_add_dois}
\alias{bib_add_dois}
\title{Add DOIs to bibliography}
\usage{
bib_add_dois(bib, strict = TRUE)
}
\arguments{
\item{bib}{the bib table}

\item{strict}{Should there be a single exact match for title and journal, if FALSE, gives the best match}
}
\value{
the bib table with updated DOIs
}
\description{
Add DOIs to bibliography
}
