% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{bibsearch}
\alias{bibsearch}
\title{Search for biblio info}
\usage{
bibsearch(title, source = NA, authors = NA, strict = TRUE)
}
\arguments{
\item{title}{The title of the work}

\item{source}{The source (journal or book)}

\item{authors}{The authors}

\item{strict}{Whether to return NULL or the best match if there isn't a single match}
}
\value{
A data frame with citation info
}
\description{
Search for biblio info
}
\examples{
\dontrun{
  bibsearch("Sample Size Justification", "Collabra Psychology")
}
}
