% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{bibtex_add_dois}
\alias{bibtex_add_dois}
\title{Add DOIs to a bib file}
\usage{
bibtex_add_dois(bibfile, save_to = NULL, strict = TRUE)
}
\arguments{
\item{bibfile}{The file path to the .bib file}

\item{save_to}{The file to save the results to; if NULL, saves to bibfile name with _doi appended}

\item{strict}{Should there be a single exact match for title and journal, if FALSE, gives the best match}
}
\value{
a bib table in the bib2df format
}
\description{
Uses OpenAlex to search for items that match the title and journal of bibtex entries that don't have a DOI and adds them in.
}
