% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_cc.R
\name{cc}
\alias{cc}
\alias{cc_metrics}
\alias{cc_compliance_metrics}
\title{Diagnose Creative Commons Licensing}
\usage{
cc_metrics(cc_license_check = NULL)

cc_compliance_metrics(cc_license_check = NULL)
}
\arguments{
\item{cc_license_check}{tibble, result from \code{\link[=license_check]{license_check()}}}
}
\description{
Diagnose Creative Commons Licensing
}
\section{Functions}{
\itemize{
\item \code{cc_metrics}: Presents normalised CC licence variants.
For example, \code{BY} (absolute and relative).

\item \code{cc_compliance_metrics}: Presents the result of the CC compliance check.
Absolute and relative normalised CC licence variants.
}}

\examples{
# obtain metadata from Crossref API
req <- get_cr_md(doi_examples$good)

# check article-level compliance
out <- cr_compliance_overview(req)
# obtain CC variants metrics
cc_metrics(out$cc_license_check)
# Obtain cc compliance check resutls metrics
cc_compliance_metrics(out$cc_license_check)
}
\seealso{
\code{\link[=license_check]{license_check()}}

Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_tdm_df}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{mc_predicates}},
\code{\link{vor_issue}()}
}
\concept{transform}
\keyword{internal}
