% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_cc.R
\name{cc_compliance_metrics}
\alias{cc_compliance_metrics}
\title{CC compliance metrics overview}
\usage{
cc_compliance_metrics(cc_license_check = NULL)
}
\arguments{
\item{cc_license_check}{tibble, result from \code{\link[=license_check]{license_check()}}}
}
\description{
Presents the result of the CC compliance check (absolute and relative)
}
\examples{
\dontrun{
# Workflow:
# First, obtain metadata from Crossref API
req <- get_cr_md(tu_dois()[7:10])

# Then, check article-level compliance
out <- cr_compliance_overview(req)

# Obtain CC compliance check resutls metrics
cc_compliance_metrics(out$cc_license_check)
}
}
\seealso{
\code{\link[=license_check]{license_check()}}
}
\keyword{internal}
