% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_cc.R
\name{cc_metrics}
\alias{cc_metrics}
\title{CC license variants metrics overview}
\usage{
cc_metrics(cc_license_check = NULL)
}
\arguments{
\item{cc_license_check}{tibble, result from \code{\link[=license_check]{license_check()}}}
}
\description{
Presents normalised CC licence variants like BY (absolute and relative)
}
