% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cermine.R
\name{cermine2bib}
\alias{cermine2bib}
\title{Parse cermine XML bib format to bibtex}
\usage{
cermine2bib(ref)
}
\arguments{
\item{ref}{the biblStruct xml object}
}
\value{
a bibentry
}
\description{
Parse cermine XML bib format to bibtex
}
\keyword{internal}
