% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_check.R
\name{code_abs_path}
\alias{code_abs_path}
\title{Return Absolute Paths}
\usage{
code_abs_path(code_text)
}
\arguments{
\item{code_text}{the text of the code, excluding comments}
}
\value{
a vector of absolute paths
}
\description{
Check code for the presence of absolute paths
}
\examples{
code_text <- c(
  "file <- 'C:/User/lakens/file.R'",
  "tmp <- '/User/lakens/file.html'",
  "convert(file, tmp)"
)
code_abs_path(code_text)
}
