% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_check.R
\name{code_extract_r}
\alias{code_extract_r}
\title{Convert Rmd/qmd files to R code only}
\usage{
code_extract_r(
  file_path = NULL,
  save_path = NULL,
  documentation = 0,
  text = NULL
)
}
\arguments{
\item{file_path}{a vector of file paths to check}

\item{save_path}{if NULL, returns a text vector, else a path to save to}

\item{documentation}{0:2 value to pass to knitr::purl}

\item{text}{alternative to file_path, pass text directly}
}
\value{
a character vector
}
\description{
Convert Rmd/qmd files to R code only
}
\examples{
file_path <- demofile("qmd")
code_text <- code_extract_r(file_path)
}
