% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_check.R
\name{code_library_lines}
\alias{code_library_lines}
\title{Get Code Library Lines}
\usage{
code_library_lines(
  code_text,
  lang = c("R", "Python", "SPSS", "SAS", "Stata", "Mplus", "MATLAB")
)
}
\arguments{
\item{code_text}{the code text for a single file}

\item{lang}{the language (we only currently handle R, Python, SPSS, SAS,
Stata, Mplus, MATLAB)}
}
\value{
a data frame with columns \code{code} and \code{line} (the line numbers on which library calls exist, after removing blank lines and comments)
}
\description{
Returns the lines on which library/require calls exist. This is a helper function for the code_check module.
}
\examples{
code_text <- c(
  "library(dplyr)",
  "",
  "# this line won't count",
  "library(tidyr)",
  "renv::install('metacheck')"
)
code_library_lines(code_text, "R")
}
