% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_check.R
\name{code_parse_r}
\alias{code_parse_r}
\title{Parse code to check for errors}
\usage{
code_parse_r(file_path = "", text = NULL)
}
\arguments{
\item{file_path}{a vector of file paths to check}

\item{text}{alternative to file_path, pass text directly}
}
\value{
a data frame with columns \code{file_path} and \code{line}
}
\description{
Parse code to check for errors
}
\examples{
file_path <- demofile("qmd")
code_parse_r(file_path)
}
