% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_check.R
\name{code_read}
\alias{code_read}
\title{Read code from files}
\usage{
code_read(file_path)
}
\arguments{
\item{file_path}{a file path or url to read in}
}
\value{
a character vector of the file contents
}
\description{
Read code from files
}
\examples{
file_path <- demofile("json")
text <- code_read(file_path)
}
