% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_check.R
\name{code_remove_comments}
\alias{code_remove_comments}
\title{Remove comments from code text}
\usage{
code_remove_comments(
  code_text,
  lang = c("R", "Python", "SPSS", "SAS", "Stata", "Mplus", "MATLAB")
)
}
\arguments{
\item{code_text}{the code text for a single file}

\item{lang}{the language (we only currently handle R, Python, SPSS, SAS,
Stata, Mplus, MATLAB)}
}
\value{
the code_text minus comment lines
}
\description{
Remove comments from code text
}
\examples{
code_text <- c(
  "# this is a comment",
  "",
  "x <- 'And this is code'"
)
code_text_nc <- code_remove_comments(code_text, "R")
}
