% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concat_tables}
\alias{concat_tables}
\title{Concatenate tables}
\usage{
concat_tables(papers, name_path)
}
\arguments{
\item{papers}{a list of paper objects}

\item{name_path}{a vector of names that get you to the table}
}
\value{
a merged table
}
\description{
Concatenate tables across a list of paper objects
}
\examples{
references <- concat_tables(psychsci[1:10], "references")
}
