% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert documents}
\usage{
convert(
  file_path,
  save_path = ".",
  method = c("auto", "bibr", "grobid", "xml"),
  ...
)
}
\arguments{
\item{file_path}{Path to the document file, or a directory of documents}

\item{save_path}{Path to a directory in which to save the JSON file}

\item{method}{whether to use bibr, grobid, or grobid_to_bibr to convert a file (see Details)}

\item{...}{further arguments to pass to convert_bibr, convert_grobid, or grobid_to_bibr}
}
\value{
the path to the JSON file
}
\description{
Uses grobid or bibr to convert a file to paper format.
}
\details{
Both bibr and grobid can handle PDF files. Only bibr can convert doc or docx files. Already-converted grobid XML files can be converted to bibr format (set crossref_lookup=TRUE to add a bib_match table). If the file_path is a directory, the method will be xml if any XML files are present, and bibr if only doc or docx files are present.
}
