% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_overview.R
\name{cr_compliance_overview}
\alias{cr_compliance_overview}
\title{Get compliance overview}
\usage{
cr_compliance_overview(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Get compliance overview
}
