% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_funder.R
\name{cr_funder_df}
\alias{cr_funder_df}
\title{Get funder list from Crossref metadata}
\usage{
cr_funder_df(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Get funder list from Crossref metadata
}
