% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_overview.R
\name{cr_has_orcid}
\alias{cr_has_orcid}
\title{Check for ORCIDs}
\usage{
cr_has_orcid(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Check for ORCIDs
}
