% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_ind.R
\name{cr_md_ind}
\alias{cr_md_ind}
\title{Other types of relevant metadata}
\usage{
cr_md_ind(cr, .group)
}
\arguments{
\item{cr}{crossref metadata}

\item{.group}{group by variable, like publisher}
}
\description{
Other types of relevant metadata
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_license_ind}()},
\code{\link{cr_tdm_df}()},
\code{\link{cr_tdm_ind}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{vor_issue}()}
}
\concept{transform}
