% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_tdm.R
\name{cr_tdm_df}
\alias{cr_tdm_df}
\title{Get text-mining links from Crossref metadata}
\usage{
cr_tdm_df(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Get text-mining links from Crossref metadata
}
