% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_ind.R
\name{cr_tdm_ind}
\alias{cr_tdm_ind}
\title{TDM overview}
\usage{
cr_tdm_ind(cr)
}
\arguments{
\item{cr}{crossref metadata}
}
\description{
TDM overview
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_license_ind}()},
\code{\link{cr_md_ind}()},
\code{\link{cr_tdm_df}()},
\code{\link{get_compliant_cc}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{license_normalise}()},
\code{\link{license_report}()},
\code{\link{license_val}()},
\code{\link{open_apc_get_}()},
\code{\link{open_apc_get}()},
\code{\link{vor_issue}()}
}
\concept{transform}
