% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacheck_app.R
\name{create_report_app}
\alias{create_report_app}
\title{Launch Create Report App}
\usage{
create_report_app(quiet = FALSE, ...)
}
\arguments{
\item{quiet}{whether to show debugging messages in the console}

\item{...}{arguments to pass to shiny::runApp}
}
\value{
NULL (invisibly)
}
\description{
Launch the Create Report app: upload a PDF and generate a report with one
click, with privacy options for what is sent to external servers.
}
\examples{
\dontrun{
create_report_app()
}

}
