% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{crossref}
\alias{crossref}
\title{Crossref info}
\usage{
crossref(doi)
}
\arguments{
\item{doi}{the DOI of the paper to get info for}
}
\value{
crossref data
}
\description{
Crossref info
}
\examples{
doi <- "10.7717/peerj.4375"
\dontrun{
  cr_info <- crossref(doi)
}
}
