% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_helpers.R
\name{data_check_constant}
\alias{data_check_constant}
\title{Flag a constant or near-constant column}
\usage{
data_check_constant(x, threshold = 0.99)
}
\arguments{
\item{x}{a vector}

\item{threshold}{if the most common non-NA value covers at least this
fraction, the column is near-constant}
}
\value{
list(problem, message, values)
}
\description{
Flag a constant or near-constant column
}
\keyword{internal}
