% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_helpers.R
\name{data_check_miscoded_missing}
\alias{data_check_miscoded_missing}
\title{Flag values that look like miscoded missing data}
\usage{
data_check_miscoded_missing(x, sentinels = .data_missing_sentinels)
}
\arguments{
\item{x}{a numeric vector}

\item{sentinels}{candidate sentinel values}
}
\value{
list(problem, message, values)
}
\description{
Flag values that look like miscoded missing data
}
\keyword{internal}
