% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_helpers.R
\name{data_check_sparse_levels}
\alias{data_check_sparse_levels}
\title{Flag sparsely represented categorical levels}
\usage{
data_check_sparse_levels(x, min_n = 2L)
}
\arguments{
\item{x}{a character or factor vector}

\item{min_n}{levels with fewer than this many observations are flagged}
}
\value{
list(problem, message, values)
}
\description{
Flag sparsely represented categorical levels
}
\keyword{internal}
