% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{datacite_doi}
\alias{datacite_doi}
\title{Doi.org Info from DataCite}
\usage{
datacite_doi(doi)
}
\arguments{
\item{doi}{the DOI(s) to get info for}
}
\value{
bib_match data frame
}
\description{
Doi.org Info from DataCite
}
\examples{
doi <- "10.5281/zenodo.2669586"
\dontrun{
doi_info <- datacite_doi(doi)
}
}
