% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-dataone.R
\name{dataone_info}
\alias{dataone_info}
\title{Retrieve info from DataONE by URL}
\usage{
dataone_info(dataone_url, id_col = 1, pb = NULL, cache = FALSE)
}
\arguments{
\item{dataone_url}{a DataONE URL or DOI, or a table containing them (e.g.,
as created by \code{\link[=dataone_links]{dataone_links()}})}

\item{id_col}{the index or name of the column that contains DataONE URLs,
if \code{dataone_url} is a table}

\item{pb}{a progress bar passed from another function}

\item{cache}{if \code{TRUE}, reuse a previously cached listing for a dataset
already looked up (see \code{\link[=repo_info_cache]{repo_info_cache()}}) instead of re-querying the
host's API. Off by default.}
}
\value{
a data frame of information
}
\description{
Retrieve info from DataONE by URL
}
\examples{
\dontrun{
  dataone_info("https://doi.org/10.18739/A2GT5FG86")
}
}
