% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{demofile}
\alias{demofile}
\title{Get a demo file}
\usage{
demofile(ext = c("json", "pdf", "docx", "doc", "xml", "qmd"))
}
\arguments{
\item{ext}{the extension of the file}
}
\value{
file path
}
\description{
Return the file path for various versions of the demo paper. Use \code{demopaper()} to directly read it as a paper object from the json file.
}
\examples{
json <- demofile()
pdf <- demofile("pdf")
}
