% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doi_examples}
\alias{doi_examples}
\title{DOIs included in the package for testing and illustration}
\format{
A list of \code{\link[biblids:doi]{biblids::doi()}} vectors, including:
\itemize{
\item \code{good}: Guaranteed to work (all are \code{\link[=is_metacheckable]{is_metacheckable()}}).
\item \code{many}: Long list of DOIs.
\item \code{weird}: Troublesome DOIs.
\item \code{biblids}: \code{\link[biblids:doi_examples]{biblids::doi_examples()}}
}
}
\usage{
doi_examples
}
\description{
DOIs included in the package for testing and illustration
}
\examples{
doi_examples$good

}
\concept{data}
\keyword{datasets}
