% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_lookup}
\alias{doi_lookup}
\title{Doi.org Info from DOI}
\usage{
doi_lookup(doi)
}
\arguments{
\item{doi}{the DOI(s) to get info for}
}
\value{
data frame with DOIs and info
}
\description{
Doi.org Info from DOI
}
\examples{
doi <- "10.7717/peerj.4375"
\dontrun{
doi_info <- doi_lookup(doi)
}
}
