% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibr.R
\name{.bibr_isalive}
\alias{.bibr_isalive}
\title{Check bibr server status}
\usage{
.bibr_isalive(api_url, api_key = Sys.getenv("SCIVRS_API_KEY"), error = TRUE)
}
\arguments{
\item{api_url}{the URL to the bibr server}

\item{api_key}{the API key to use (NULL if local)}

\item{error}{whether to generate and error on failure}
}
\value{
boolean
}
\description{
Check bibr server status
}
\keyword{internal}
