% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-crossref.R
\name{.bibtype_convert}
\alias{.bibtype_convert}
\title{Convert crossref/doi types to bibtex types}
\usage{
.bibtype_convert(type)
}
\arguments{
\item{type}{a vector of crossref types}
}
\value{
a vector of bibtext types
}
\description{
Convert crossref/doi types to bibtex types
}
\examples{
crossref_types <- c("book-part",
                    "journal-article",
                    "monograph",
                    NA,
                    "unmatched-type")
.bibtype_convert(crossref_types)
}
\keyword{internal}
