% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{.crossref_batch_query}
\alias{.crossref_batch_query}
\title{Batch query crossref}
\usage{
.crossref_batch_query(urls, batch_size = 5)
}
\arguments{
\item{urls}{A vector of URLs}

\item{batch_size}{Size of each batch}
}
\value{
a list of responses
}
\description{
Batch query crossref
}
\keyword{internal}
