% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{.crossref_parse_item}
\alias{.crossref_parse_item}
\title{Parse a CrossRef item into a data frame row}
\usage{
.crossref_parse_item(
  item,
  select = c("DOI", "type", "title", "author", "container-title", "volume", "issue",
    "page", "URL", "abstract", "year", "error")
)
}
\arguments{
\item{item}{a list from CrossRef API response}

\item{select}{fields to select}
}
\value{
a data frame
}
\description{
Parse a CrossRef item into a data frame row
}
\keyword{internal}
