% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{.crossref_query_parse}
\alias{.crossref_query_parse}
\title{Parse crossref query items into a table}
\usage{
.crossref_query_parse(items, min_score, select)
}
\arguments{
\item{items}{list of items from CrossRef query response}

\item{min_score}{minimum score threshold}

\item{select}{fields to select}
}
\value{
a data frame
}
\description{
Parse crossref query items into a table
}
\keyword{internal}
