% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-dataone.R
\name{.dataone_host}
\alias{.dataone_host}
\title{Identify which known DataONE member node a URL or DOI belongs to}
\usage{
.dataone_host(dataone_url)
}
\arguments{
\item{dataone_url}{a vector of URLs or DOIs}
}
\value{
a character vector of hosts (see .dataone_hosts()), or \code{NA} for
one that matches no known host
}
\description{
Identify which known DataONE member node a URL or DOI belongs to
}
\keyword{internal}
