% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-dataone.R
\name{.dataone_info}
\alias{.dataone_info}
\title{Retrieve info from one DataONE object}
\usage{
.dataone_info(pid, host, pb = NULL)
}
\arguments{
\item{pid}{a DataONE persistent identifier (e.g. \code{"doi:10.18739/..."})}

\item{host}{the member node host to query (see .dataone_hosts())}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from one DataONE object
}
\keyword{internal}
