% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-dataone.R
\name{.dataone_pid}
\alias{.dataone_pid}
\title{Get a DataONE persistent identifier (PID) from a URL or DOI}
\usage{
.dataone_pid(dataone_url)
}
\arguments{
\item{dataone_url}{a vector of URLs or DOIs to DataONE-hosted objects}
}
\value{
a character vector of PIDs, normalised to DataONE's own
\verb{doi:<prefix>/<suffix>} form
}
\description{
Get a DataONE persistent identifier (PID) from a URL or DOI
}
\keyword{internal}
