% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-dataverse.R
\name{.dataverse_info}
\alias{.dataverse_info}
\title{Retrieve info from one Dataverse dataset}
\usage{
.dataverse_info(host, doi, pb = NULL)
}
\arguments{
\item{host}{the Dataverse installation's hostname (e.g. "dataverse.harvard.edu")}

\item{doi}{the dataset's persistentId (a DOI, e.g. "10.7910/DVN/GGDUND")}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from one Dataverse dataset
}
\keyword{internal}
