% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-dryad.R
\name{.dryad_doi}
\alias{.dryad_doi}
\title{Get Dryad DOI from URL}
\usage{
.dryad_doi(dryad_url)
}
\arguments{
\item{dryad_url}{a vector of URLs or DOIs to dryad datasets}
}
\value{
a character vector of DOIs in "10.5061/dryad.xxxxx" form
}
\description{
Get Dryad DOI from URL
}
\keyword{internal}
