% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-dryad.R
\name{.dryad_info}
\alias{.dryad_info}
\title{Retrieve info from one Dryad dataset}
\usage{
.dryad_info(dryad_doi, pb = NULL)
}
\arguments{
\item{dryad_doi}{the dataset's DOI (e.g. "10.5061/dryad.j1fd7")}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from one Dryad dataset
}
\keyword{internal}
