% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-figshare.R
\name{.figshare_id}
\alias{.figshare_id}
\title{Get Figshare article ID from URL or DOI}
\usage{
.figshare_id(figshare_url)
}
\arguments{
\item{figshare_url}{a vector of URLs or DOIs to figshare articles}
}
\value{
a character vector of numeric article IDs
}
\description{
Get Figshare article ID from URL or DOI
}
\keyword{internal}
