% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-figshare.R
\name{.figshare_info}
\alias{.figshare_info}
\title{Retrieve info from one Figshare article}
\usage{
.figshare_info(figshare_id, host = "api.figshare.com", pb = NULL)
}
\arguments{
\item{figshare_id}{a Figshare article ID}

\item{host}{the Figshare-compatible API host to query}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from one Figshare article
}
\keyword{internal}
