% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-fsd.R
\name{.fsd_info}
\alias{.fsd_info}
\title{Retrieve info from FSD by URL}
\usage{
.fsd_info(fsd_url, pb = NULL)
}
\arguments{
\item{fsd_url}{an FSD URL}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from FSD by URL
}
\keyword{internal}
