% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobid.R
\name{.grobid_isalive}
\alias{.grobid_isalive}
\title{Check grobid server status}
\usage{
.grobid_isalive(api_url, error = TRUE)
}
\arguments{
\item{api_url}{the URL to the grobid server}

\item{error}{whether to generate and error on failure}
}
\value{
boolean
}
\description{
Check grobid server status
}
\keyword{internal}
