% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{.llm_model_list_groq}
\alias{.llm_model_list_groq}
\title{Get Groq Models}
\usage{
.llm_model_list_groq()
}
\value{
a table of model info
}
\description{
Returns a list of available models in groq, excluding whisper and vision models (for audio and images) See \url{https://console.groq.com/docs/models} for more information.
}
\keyword{internal}
