% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-mendeley.R
\name{.mendeley_id}
\alias{.mendeley_id}
\title{Get Mendeley Data dataset ID from URL or DOI}
\usage{
.mendeley_id(mendeley_url)
}
\arguments{
\item{mendeley_url}{a vector of URLs or DOIs to Mendeley Data datasets}
}
\value{
a character vector of dataset IDs
}
\description{
Get Mendeley Data dataset ID from URL or DOI
}
\keyword{internal}
