% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-mendeley.R
\name{.mendeley_info}
\alias{.mendeley_info}
\title{Retrieve info from one Mendeley Data dataset}
\usage{
.mendeley_info(mendeley_id, pb = NULL)
}
\arguments{
\item{mendeley_id}{a Mendeley Data dataset ID}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from one Mendeley Data dataset
}
\keyword{internal}
