% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{.openalex_add_abstract}
\alias{.openalex_add_abstract}
\title{Add abstract from inverted index}
\usage{
.openalex_add_abstract(info)
}
\arguments{
\item{info}{OpenAlex response list}
}
\value{
the info list with abstract field added
}
\description{
Add abstract from inverted index
}
\keyword{internal}
