% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-osf-metadata.R
\name{.osf_meta_dir}
\alias{.osf_meta_dir}
\title{Name of the folder metadata is written to inside a project's download folder}
\usage{
.osf_meta_dir
}
\description{
Name of the folder metadata is written to inside a project's download folder
}
\keyword{internal}
