% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-osf-helpers.R
\name{.osf_pat_validate}
\alias{.osf_pat_validate}
\title{OSF PAT Validation}
\usage{
.osf_pat_validate(osf_pat = Sys.getenv("OSF_PAT"))
}
\arguments{
\item{osf_pat}{the OSF PAT (read from renviron by default)}
}
\value{
logical (TRUE if OSF_PAT is set and valid)
}
\description{
Checks for validity of the OSF PAT and unsets it if needed.
}
\keyword{internal}
