% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{.paper_coerce}
\alias{.paper_coerce}
\title{Coerce paper object types}
\usage{
.paper_coerce(paper)
}
\arguments{
\item{paper}{a paper object}
}
\value{
a paper object
}
\description{
Convert columns to the correct type
}
\keyword{internal}
