% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-psycharchives.R
\name{.psycharchives_info}
\alias{.psycharchives_info}
\title{Retrieve info from PsychArchives by URL}
\usage{
.psycharchives_info(pa_url, pb = NULL)
}
\arguments{
\item{pa_url}{a PsychArchives URL}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from PsychArchives by URL
}
\keyword{internal}
