% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-bibr.R
\name{.read_bibr}
\alias{.read_bibr}
\title{Read bibr JSON file}
\usage{
.read_bibr(file_path, include_images = FALSE)
}
\arguments{
\item{file_path}{path to the JSON file}

\item{include_images}{whether to include images in the figures table of the paper object (they make object size larger)}
}
\value{
a paper object
}
\description{
Read bibr JSON file
}
\keyword{internal}
