% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibr.R
\name{.read_bibr_zip}
\alias{.read_bibr_zip}
\title{Read legacy bibr ZIP file (Arrow format)}
\usage{
.read_bibr_zip(file_path)
}
\arguments{
\item{file_path}{path to the zip file}
}
\value{
a paper object
}
\description{
Read legacy bibr ZIP file (Arrow format)
}
\keyword{internal}
