% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.report_title_slug}
\alias{.report_title_slug}
\title{Report from module output}
\usage{
.report_title_slug(title)
}
\arguments{
\item{module_output}{the output of a \code{module_run()}}

\item{header}{header level (default 2)}
}
\value{
text
}
\description{
Report from module output
}
\examples{
paper <- demopaper()
op <- module_run(paper, "stat_p_exact")
module_report(op) |> cat()
}
