% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-reshare.R
\name{.reshare_id}
\alias{.reshare_id}
\title{Get ReShare eprint ID from URL or DOI}
\usage{
.reshare_id(reshare_url)
}
\arguments{
\item{reshare_url}{a vector of URLs or DOIs to reshare deposits}
}
\value{
a character vector of numeric eprint IDs
}
\description{
Get ReShare eprint ID from URL or DOI
}
\keyword{internal}
