% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-reshare.R
\name{.reshare_info}
\alias{.reshare_info}
\title{Retrieve info from one ReShare deposit}
\usage{
.reshare_info(reshare_id, pb = NULL)
}
\arguments{
\item{reshare_id}{a ReShare eprint ID}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from one ReShare deposit
}
\keyword{internal}
