% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobid.R
\name{.tei_authors}
\alias{.tei_authors}
\title{Get authors from TEI type XML}
\usage{
.tei_authors(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
authors table
}
\description{
Get authors from TEI type XML
}
\keyword{internal}
