% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobid.R
\name{.tei_text}
\alias{.tei_text}
\title{Get full text from TEI type XML}
\usage{
.tei_text(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
a data frame with all text
}
\description{
Get full text from TEI type XML
}
\keyword{internal}
