% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobid.R
\name{.tei_url}
\alias{.tei_url}
\title{Get URLs from text table}
\usage{
.tei_url(text_table)
}
\arguments{
\item{text_table}{The text table for the paper}
}
\value{
xrefs table
}
\description{
Get URLs from text table
}
\keyword{internal}
