% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobid.R
\name{.tei_xrefs}
\alias{.tei_xrefs}
\title{Get cross references from TEI type XML}
\usage{
.tei_xrefs(xml, text_table)
}
\arguments{
\item{xml}{The XML}

\item{text_table}{The text table for the paper}
}
\value{
xrefs table
}
\description{
Get cross references from TEI type XML
}
\keyword{internal}
