% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svutils-xml.R
\name{.xml_find1_text}
\alias{.xml_find1_text}
\title{Find and return first text from XML by xpath}
\usage{
.xml_find1_text(xml, xpath)
}
\arguments{
\item{xml}{the xml document, node, or nodeset}

\item{xpath}{a string containing an xpath expression}
}
\value{
text
}
\description{
Find and return first text from XML by xpath
}
\keyword{internal}
