% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svutils-xml.R
\name{.xml_find_text}
\alias{.xml_find_text}
\title{Find and return text from XML by xpath}
\usage{
.xml_find_text(xml, xpath, join = NULL)
}
\arguments{
\item{xml}{the xml document, node, or nodeset}

\item{xpath}{a string containing an xpath expression}

\item{join}{optional string to join vectors}
}
\value{
text
}
\description{
This function trims and replaces multiple spaces
}
\keyword{internal}
