% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svutils-xml.R
\name{.xml_read_grobid}
\alias{.xml_read_grobid}
\title{Read in a grobid XML file}
\usage{
.xml_read_grobid(path)
}
\arguments{
\item{path}{Path to a grobid/TEI XML file}
}
\value{
a list with class "xml_document" for xml2
}
\description{
Read in, strip TEI namespace, and fix common problems.
}
\examples{
path <- demofile("xml")
xml <- .xml_read_grobid(path)
}
\keyword{internal}
