% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-zenodo-upload.R
\name{.zenodo_api}
\alias{.zenodo_api}
\title{Zenodo API base URL}
\usage{
.zenodo_api(sandbox = TRUE)
}
\arguments{
\item{sandbox}{whether to use the sandbox server}
}
\value{
the base API URL
}
\description{
Zenodo API base URL
}
\keyword{internal}
