% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-zenodo-upload.R
\name{.zenodo_auth}
\alias{.zenodo_auth}
\title{Add Zenodo auth to a request}
\usage{
.zenodo_auth(req, token)
}
\arguments{
\item{req}{an httr2 request object}

\item{token}{the Zenodo personal access token}
}
\value{
the modified request
}
\description{
Add Zenodo auth to a request
}
\keyword{internal}
