% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-zenodo.R
\name{.zenodo_id}
\alias{.zenodo_id}
\title{Get zenodo ID from URL}
\usage{
.zenodo_id(zenodo_url)
}
\arguments{
\item{zenodo_url}{a vector of URLs to zenodo repos}
}
\value{
a character vector of IDs
}
\description{
Get zenodo ID from URL
}
\keyword{internal}
