% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-zenodo.R
\name{.zenodo_info}
\alias{.zenodo_info}
\title{Retrieve info from Zenodo by ID}
\usage{
.zenodo_info(zenodo_id, pb = NULL)
}
\arguments{
\item{zenodo_id}{a Zenodo ID or URL}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from Zenodo by ID
}
\keyword{internal}
