% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-figshare.R
\name{figshare_info}
\alias{figshare_info}
\title{Retrieve info from Figshare by URL}
\usage{
figshare_info(
  figshare_url,
  id_col = 1,
  host = "api.figshare.com",
  pb = NULL,
  cache = FALSE
)
}
\arguments{
\item{figshare_url}{a Figshare URL or DOI, or a table containing them
(e.g., as created by \code{\link[=figshare_links]{figshare_links()}})}

\item{id_col}{the index or name of the column that contains Figshare URLs,
if \code{figshare_url} is a table}

\item{host}{the Figshare-compatible API host to query. Default
\code{"api.figshare.com"}; \code{\link[=researchdata4tu_info]{researchdata4tu_info()}} calls this with
\code{"data.4tu.nl"} instead (see the note at the top of this file).}

\item{pb}{a progress bar passed from another function}

\item{cache}{if \code{TRUE}, reuse a previously cached listing for an article
already looked up (see \code{\link[=repo_info_cache]{repo_info_cache()}}) instead of re-querying the
API. Off by default.}
}
\value{
a data frame of information
}
\description{
Retrieve info from Figshare by URL
}
\examples{
\dontrun{
  figshare_info("https://doi.org/10.6084/m9.figshare.18093368.v1")
}
}
