% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{file_category}
\alias{file_category}
\title{Categorise files}
\usage{
file_category(contents)
}
\arguments{
\item{contents}{a table with columns name, path such as from \code{osf_contents()}}
}
\value{
the table with new column file_category
}
\description{
Categorise files
}
\examples{
contents <- c("script.R", "data.csv", "README", "codebook.csv")
file_category(contents)
}
