% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{filetype}
\alias{filetype}
\title{Get file Type from Extension}
\usage{
filetype(filename)
}
\arguments{
\item{filename}{the file name}
}
\value{
a named vector of file types
}
\description{
Get file Type from Extension
}
\examples{
filetype("script.R")
}
