% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_database.R
\name{FLoRA}
\alias{FLoRA}
\title{FORRT Replication Database (FLoRA)}
\format{
A data frame with 8 columns:
\describe{
\item{doi_o}{DOI of original study}
\item{apa_ref_o}{APA reference of original study}
\item{doi_r}{DOI of replication study (may be NA if url_r is provided)}
\item{apa_ref_r}{APA reference of replication study}
\item{url_r}{URL of replication study (used when DOI is not available)}
\item{outcome}{replication outcome}
\item{outcome_quote}{quote describing replication outcome}
\item{type}{type of replication}
}
}
\source{
\url{https://osf.io/9r62x/files/t4j8f}
}
\usage{
FLoRA()
}
\value{
a data frame
}
\description{
FLoRA database containing DOIs of original studies and replications. Use \code{FLoRA_date()} to find the date it was downloaded, and \code{FLoRA_update()} to update it.
}
\examples{
FLoRA()
}
