% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_ind.R
\name{gather_ind_table}
\alias{gather_ind_table}
\title{Create compliance overview table}
\usage{
gather_ind_table(cr)
}
\arguments{
\item{cr}{crossref metadata}
}
\description{
Create compliance overview table
}
