% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cermine.R
\name{get_cermine_authors}
\alias{get_cermine_authors}
\title{Get author info from CermineXML}
\usage{
get_cermine_authors(xml)
}
\arguments{
\item{xml}{The grobid XML}
}
\value{
an author list
}
\description{
Get author info from CermineXML
}
\keyword{internal}
