% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cermine.R
\name{get_cermine_full_text}
\alias{get_cermine_full_text}
\title{Add section info to full text table}
\usage{
get_cermine_full_text(xml, id = "")
}
\arguments{
\item{xml}{The cermine XML}

\item{id}{An ID for the paper}
}
\value{
a data frame of the classified full text
}
\description{
Add section info to full text table
}
\keyword{internal}
