% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cermine.R
\name{get_cermine_refs}
\alias{get_cermine_refs}
\title{Get references from cermine XML}
\usage{
get_cermine_refs(xml)
}
\arguments{
\item{xml}{The cermine XML}
}
\value{
a list with a data frame of references and a data frame of citation sentences
}
\description{
Get references from cermine XML
}
\keyword{internal}
