% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_fetch.R, R/import.R
\name{get_cr_md}
\alias{get_cr_md}
\title{Get Crossref metadata from API}
\usage{
get_cr_md(dois, .progress = "text")

get_cr_md(dois, .progress = "text")
}
\arguments{
\item{dois}{character vector with DOIs}

\item{.progress}{show progress bar, use "none" if no progress should be displayed}
}
\description{
Get Crossref metadata from API

Get Crossref metadata from API
}
\concept{ETL import}
