% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_cc.R
\name{get_delayed_license}
\alias{get_delayed_license}
\title{Extract records with license delay}
\usage{
get_delayed_license(license_df)
}
\arguments{
\item{license_df}{normalized license metadata from \code{\link[=get_license_md]{get_license_md()}}}
}
\description{
Extract records with license delay
}
\details{
Obtain records with CC license where license start_date
and earliest publication date for the version of records
differs, suggesting delayed CC license provision.
}
