% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_cc.R
\name{get_license_md}
\alias{get_license_md}
\title{Extract license info from Crossref metadata}
\usage{
get_license_md(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Extract license info from Crossref metadata
}
