% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grobid.R
\name{get_submission}
\alias{get_submission}
\title{Get submission info from XML}
\usage{
get_submission(xml)
}
\arguments{
\item{xml}{The grobid XML}
}
\value{
a submission
}
\description{
Get submission info from XML
}
\keyword{internal}
