% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grobid.R
\name{get_tables}
\alias{get_tables}
\title{Get Tables from Grobid XML}
\usage{
get_tables(xml)
}
\arguments{
\item{xml}{The grobid XML}
}
\value{
a list of tables
}
\description{
Get Tables from Grobid XML
}
\keyword{internal}
