% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_apc_fetch.R
\name{gh_file}
\alias{gh_file}
\title{Gets a file from a github repo, using the Data API blob endpoint}
\usage{
gh_file(
  url = NULL,
  ref = NULL,
  owner = NULL,
  repo = NULL,
  path = NULL,
  to_disk = TRUE,
  destfile = NULL,
  .token = NULL,
  .api_url = NULL,
  .method = "GET",
  .limit = NULL,
  .send_headers = NULL
)
}
\arguments{
\item{url}{the URL of the file to download via API, of the form
`github.com/:owner/:repo/blob/:path}

\item{ref}{the reference of a commit: a branch name, tag, or commit SHA}

\item{owner, repo, path, ref}{alternate way to specify the file.  These will
override values in \code{url}}

\item{to_disk, destfile}{write file to disk (default=TRUE)?  If so, use the
name in \code{destfile}, or the original filename by default}

\item{.token, .api_url, .method, .limit, .send_headers}{arguments passed on to
\link[gh:gh]{gh::gh}}
}
\value{
Either the local path of the downloaded file (default), or a raw
vector
}
\description{
This avoids the 1MB limit of the content API and uses \link[gh:gh]{gh::gh} to deal with
authorization and such.  See https://developer.github.com/v3/git/blobs/
}
\author{
Noam Ross
}
