% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_config}
\alias{github_config}
\title{GitHub Configuration}
\usage{
github_config(req)
}
\arguments{
\item{req}{an httr2 request object}
}
\value{
the modified request
}
\description{
Adds GitHub auth and accept headers to an httr2 request.
}
\keyword{internal}
