% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-github.R
\name{github_files}
\alias{github_files}
\title{Get File List from GitHub}
\usage{
github_files(repo, dir = "", recursive = FALSE)
}
\arguments{
\item{repo}{The URL of the repository (in the format "username/repo" or "https://github.com/username/repo")}

\item{dir}{an optional directory name to search}

\item{recursive}{whether to search the files recursively}
}
\value{
a data frame of files
}
\description{
Get File List from GitHub
}
\examples{
\dontrun{
github_files("scienceverse/metacheck")
}
}
