% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-github.R
\name{github_links}
\alias{github_links}
\title{Find GitHub Links in Papers}
\usage{
github_links(paper)
}
\arguments{
\item{paper}{a paper object or paperlist object}
}
\value{
a table with the GitHub url in the first (text) column
}
\description{
GitHub links can be in PDFs in several ways.
}
\examples{
\dontrun{
psychsci <- papers_load("psychsci", cache = TRUE)
github_links(psychsci)
}
}
