% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-gitlab.R
\name{gitlab_repo}
\alias{gitlab_repo}
\title{Get Short GitLab Project Path}
\usage{
gitlab_repo(repo)
}
\arguments{
\item{repo}{The URL of the project (in the format "namespace/project" or
"https://gitlab.com/namespace/project")}
}
\value{
character string of the "namespace/project" path, or NULL if the
project does not exist / is not reachable
}
\description{
Get Short GitLab Project Path
}
\examples{
gitlab_repo("gitlab-org/gitlab")
gitlab_repo("https://gitlab.com/gitlab-org/gitlab/")
gitlab_repo("https://gitlab.com/gitlab-org/gitlab.git")
}
