% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacheck_metrics.R
\name{indexing_status}
\alias{indexing_status}
\title{Measure DOI indexing status}
\usage{
indexing_status(dois = NULL, .md = NULL)
}
\arguments{
\item{dois}{character vector, submitted DOIs}

\item{.md}{list, returned from \code{\link[=cr_compliance_overview]{cr_compliance_overview()}}}
}
\description{
Check the extent to which DOIs are available through Crossref
}
