% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{is_paper}
\alias{is_paper}
\title{Detect a paper object}
\usage{
is_paper(paper)
}
\arguments{
\item{paper}{the object to test}
}
\value{
logical
}
\description{
Lightweight check if an object is a paper vs paperlist. Use \code{paper_validate()} for a thorough check.
}
\keyword{internal}
