% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{jats_bib}
\alias{jats_bib}
\title{Get bibliography from JATS APA-DTD or NLM DTD type XML}
\usage{
jats_bib(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
an list of two tables: references and citations
}
\description{
Get bibliography from JATS APA-DTD or NLM DTD type XML
}
\keyword{internal}
