% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{jats_xrefs}
\alias{jats_xrefs}
\title{Get cross references from JATS APA-DTD or NLM DTD type XML}
\usage{
jats_xrefs(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
xrefs table
}
\description{
Get cross references from JATS APA-DTD or NLM DTD type XML
}
\keyword{internal}
