% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_cc.R
\name{license_check}
\alias{license_check}
\title{Extract and validate Crossref license metadata}
\usage{
license_check(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Workflow to check for compliant open content license metadata in Crossref.
License metadata is considered as compliant, if a Creative Commons
license is provided for the version-of-record without delay indicated
by the license date.
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_license_ind}()},
\code{\link{cr_md_ind}()},
\code{\link{cr_tdm_df}()},
\code{\link{cr_tdm_ind}()},
\code{\link{get_license_md}()},
\code{\link{vor_issue}()}
}
\concept{transform}
