% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_fetch.R
\name{license_normalise}
\alias{license_normalise}
\title{Normalise license info from Crossref}
\usage{
license_normalise(cr)
}
\arguments{
\item{cr}{tibble with non-compliant license metadata retrieved from the Crossref API with \code{rcrossref:.cr_works()}}
}
\description{
Normalise license info from Crossref
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_license_ind}()},
\code{\link{cr_md_ind}()},
\code{\link{cr_tdm_df}()},
\code{\link{cr_tdm_ind}()},
\code{\link{get_compliant_cc}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{license_report}()},
\code{\link{license_val}()},
\code{\link{open_apc_get_}()},
\code{\link{open_apc_get}()},
\code{\link{vor_issue}()}
}
\concept{transform}
