% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_fetch.R
\name{license_report}
\alias{license_report}
\title{Obtain faulty DOIs}
\usage{
license_report(cr)
}
\arguments{
\item{cr}{tibble with non-compliant license metadata retrieved from the
Crossref API with \code{rcrossref:.cr_works()}}
}
\description{
Obtain faulty DOIs
}
