% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_fetch.R
\name{license_val}
\alias{license_val}
\title{check compliance

compliant_with_license <- license_val(cr)get non-compliant records

non_compliant_dois <-
dois\link{!tolower(dois) \%in\% tolower(compliant_with_license$doi)}determine why

if (!is.null(non_compliant_dois))
purrr::map_df(non_compliant_dois, function(x) {
dplyr::filter(cr, doi \%in\% non_compliant_dois) \%>\%
license_report()
})
}
#' License checker}
\usage{
license_val(cr)
}
\arguments{
\item{cr}{tibble with Crossref metadata retrieved from
API with \code{rcrossref:.cr_works()}}
}
\description{
check compliance

compliant_with_license <- license_val(cr)get non-compliant records

non_compliant_dois <-
dois\link{!tolower(dois) \%in\% tolower(compliant_with_license$doi)}determine why

if (!is.null(non_compliant_dois))
purrr::map_df(non_compliant_dois, function(x) {
dplyr::filter(cr, doi \%in\% non_compliant_dois) \%>\%
license_report()
})
}
#' License checker
}
