% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_model}
\alias{llm_model}
\title{Set the default LLM model}
\usage{
llm_model(model)
}
\arguments{
\item{model}{the name of the model}
}
\description{
Use \code{llm_model_list()} to get a list of available models
}
