% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_use}
\alias{llm_use}
\title{Set or get papercheck LLM use}
\usage{
llm_use(llm_use = NULL)
}
\arguments{
\item{verbose}{if logical, sets whether to use LLMs}
}
\value{
the current option value (logical)
}
\description{
Set or get papercheck LLM use
}
\examples{
llm_use()
}
