% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-local.R
\name{local_files}
\alias{local_files}
\title{List Local Files}
\usage{
local_files(path)
}
\arguments{
\item{path}{path to a local directory}
}
\value{
a data frame with columns \code{repo_url}, \code{file_name}, \code{file_url},
\code{file_location}, \code{file_size}, \code{file_type}
}
\description{
Lists all files in a local directory recursively and returns a data frame
compatible with the \code{repo_check} output table, for use with \code{code_check}.
}
\examples{
\dontrun{
  local_files("C:/my_project")
}
}
