% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logger}
\alias{logger}
\title{Log messages}
\usage{
logger(label = "", contents = list(), logpath = NULL)
}
\arguments{
\item{contents}{a named list of the log contents}

\item{logpath}{an optional file path to save the log in}

\item{context}{a string with the context (e.g.,module name)}
}
\value{
called for side effects of writing to log, returns logpath
}
\description{
Log messages
}
\examples{
logpath <- tempfile(fileext = ".log")
logger("test", list(x = 1), logpath)
jsonlite::read_json(logpath)
}
