% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logpath}
\alias{logpath}
\title{Get log path}
\usage{
logpath()
}
\arguments{
\item{module}{The name of the module (optional)}
}
\value{
the log file path
}
\description{
Checks the most recent log file created that day and creates a new one if missing or larger than 1MB.
}
\examples{
logpath()
}
\keyword{internal}
