% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\docType{data}
\name{mc_langs}
\alias{mc_langs}
\title{Available languages for metacheck}
\format{
An object of class \code{character} of length 2.
}
\usage{
mc_langs
}
\description{
Available languages for metacheck
}
\keyword{internal}
