% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{mc_long_docs}
\alias{mc_long_docs}
\alias{mc_long_docs_string}
\title{Long translations for metacheck}
\usage{
mc_long_docs(..., lang = mc_langs)

mc_long_docs_string(...)
}
\arguments{
\item{...}{character vectors, specifying subdirectory and file(s)
    within some package.  The default, none, returns the
    root of the package.  Wildcards are not supported.}

\item{lang}{Character scalar giving the anguage of the report.
\emph{Not} reactive as in \code{\link[=mcControlsServer]{mcControlsServer()}}.}
}
\description{
Longer translations are included as separate \verb{*.md} files,
and are not covered in \code{translation.json} powering \code{\link[=mc_translator]{mc_translator()}}.
}
\section{Functions}{
\itemize{
\item \code{mc_long_docs_string}: Return content of file
}}

\keyword{internal}
