% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretest.R
\name{mc_predicates}
\alias{mc_predicates}
\alias{is_in_limit}
\alias{is_doi_cr_md}
\alias{is_doi_cr_type}
\title{Metacheck custom predicates}
\usage{
is_in_limit(x, limit = getOption("mc_limit", 1000L))

is_doi_cr_md(x)

is_doi_cr_type(x, type = types_allowed)
}
\arguments{
\item{x}{A vector created by, or convertable to \code{\link[biblids:doi]{doi()}}.}

\item{limit}{Scalar integer, giving the number of DOIs to be submitted to the metacheck.}

\item{type}{Character scalar, must be one of the types from \code{\link[rcrossref:cr_types]{rcrossref::cr_types()}}.}
}
\description{
Metacheck custom predicates
}
\section{Functions}{
\itemize{
\item \code{is_in_limit}: Is the DOI in the limit?

\item \code{is_doi_cr_md}: Is there metadata for the DOI on Crossref?

\item \code{is_doi_cr_type}: Is the DOI a type (per Crossref)?
}}

\seealso{
Other transform: 
\code{\link{cc}},
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_tdm_df}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{vor_issue}()}
}
\concept{transform}
