% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communicate.R
\name{mcApp}
\alias{mcApp}
\title{Shiny webapp for metacheck}
\usage{
mcApp()
}
\description{
Shiny webapp for metacheck
}
\seealso{
Other communicate: 
\code{\link{emailReport}()},
\code{\link{mcControls}},
\code{\link{render_email}()},
\code{\link{report}},
\code{\link{runMetacheck}()},
\code{\link{smtp_send_mc}()}
}
\concept{communicate}
