% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communicate.R
\name{mcInput}
\alias{mcInput}
\alias{mcInputApp}
\alias{mcInputUI}
\alias{mcInputServer}
\title{Enter metacheck inputs through a shiny module}
\usage{
mcInputApp()

mcInputUI(id)

mcInputServer(id)
}
\arguments{
\item{id}{The id string to be namespaced (optional).}
}
\description{
Enter metacheck inputs through a shiny module
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{mcInputApp}: Test App
}
\itemize{
\item \code{mcInputUI}: Module UI
}
\itemize{
\item \code{mcInputServer}: Module server
}}}

\seealso{
Other communicate: 
\code{\link{emailReport}()},
\code{\link{render_email}()},
\code{\link{report}},
\code{\link{runMetacheck}()},
\code{\link{smtp_send_metacheck}()}
}
\concept{communicate}
