% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-mendeley.R
\name{mendeley_info}
\alias{mendeley_info}
\title{Retrieve info from Mendeley Data by URL}
\usage{
mendeley_info(mendeley_url, id_col = 1, pb = NULL, cache = FALSE)
}
\arguments{
\item{mendeley_url}{a Mendeley Data URL or DOI, or a table containing them
(e.g., as created by \code{\link[=mendeley_links]{mendeley_links()}})}

\item{id_col}{the index or name of the column that contains Mendeley Data
URLs, if \code{mendeley_url} is a table}

\item{pb}{a progress bar passed from another function}

\item{cache}{if \code{TRUE}, reuse a previously cached listing for a dataset
already looked up (see \code{\link[=repo_info_cache]{repo_info_cache()}}) instead of re-querying the
API. Off by default.}
}
\value{
a data frame of information
}
\description{
Retrieve info from Mendeley Data by URL
}
\examples{
\dontrun{
  mendeley_info("https://doi.org/10.17632/vjtxybrc28.1")
}
}
