% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacheck_app.R
\name{metacheck_app}
\alias{metacheck_app}
\title{Launch Shiny App}
\usage{
metacheck_app(paper = NULL, quiet = FALSE, ...)
}
\arguments{
\item{paper}{optional paper or paperlist to load}

\item{quiet}{whether to show the debugging messages in the console}

\item{...}{arguments to pass to shiny::runApp}
}
\value{
A paper object created or edited by the app
}
\description{
Create a meta-study file interactively in a shiny app that runs locally in RStudio or your web browser (recommended).
}
\examples{
\dontrun{
s <- metacheck_app()
}

}
