% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_overview.R
\name{metrics_overview}
\alias{metrics_overview}
\title{Metacheck metrics overview}
\usage{
metrics_overview(cr_overview = NULL)
}
\arguments{
\item{cr_overview}{tibble obtained with \code{\link[=cr_compliance_overview]{cr_compliance_overview()}}}
}
\description{
Presents the metacheck core compliance measures
}
\details{
The following metrics are presented (absolute and relative):
\itemize{
\item Creative commons license availability (\verb{CC License})
\item Compliant creative commons license availability (\verb{Compliant CC})
\item Support of links to full-texts for text and data mining (\verb{TDM Support})
\item Metadata records with links between publications and funder information (\verb{Funder info})
\item Metadata records with links between publications and ORCID  (\code{ORCID})
\item Metadata records with \href{https://i4oa.org/}{Open Abstracts}
\item Metadata records with \href{https://i4oc.org/}{Open Citations}
}
}
\examples{
# First, obtain metadata from Crossref API
req <- get_cr_md(doi_examples$good)

# Then, check article-level compliance
out <- cr_compliance_overview(req)

# Finally, obtain compliance metrics overview
metrics_overview(out$cr_overview)
}
\keyword{internal}
