% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_find}
\alias{module_find}
\title{Find a module by name or path}
\usage{
module_find(module)
}
\arguments{
\item{module}{the name of a module or path to a module}
}
\value{
the path to the module
}
\description{
Find a module by name or path
}
\keyword{internal}
