% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_info}
\alias{module_info}
\title{Get module information}
\usage{
module_info(module)
}
\arguments{
\item{module}{the name of a module or path to a module}
}
\value{
a list of module info
}
\description{
Get module information
}
\examples{
module_info("all_p_values")
}
