% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_run_llm}
\alias{module_run_llm}
\title{Run LLM module}
\usage{
module_run_llm(paper, args)
}
\arguments{
\item{paper}{the paper object (or list of object)}

\item{args}{a list of arguments to \code{llm()}}
}
\value{
data frame
}
\description{
Run LLM module
}
\keyword{internal}
