% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_template}
\alias{module_template}
\title{Create a Module from a Template}
\usage{
module_template(module_name, path = "./modules")
}
\arguments{
\item{module_name}{The short name of the module (should contain only letters, numbers, and _)}

\item{path}{The path to save the module (defaults to a directory called "modules" in the working directory)}
}
\value{
the file path (invisibly)
}
\description{
Create a Module from a Template
}
