% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{nlm_authors}
\alias{nlm_authors}
\title{Get author info from JATS NLM-DTD type XML}
\usage{
nlm_authors(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
an author list
}
\description{
Get author info from JATS NLM-DTD type XML
}
\keyword{internal}
