% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{nlm_full_text}
\alias{nlm_full_text}
\title{Get full text from JATS NLM-DTD type XML}
\usage{
nlm_full_text(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
a data frame with all text
}
\description{
Get full text from JATS NLM-DTD type XML
}
\keyword{internal}
