% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{nlm_info}
\alias{nlm_info}
\title{Get article info from JATS NLM-DTD type XML}
\usage{
nlm_info(xml, filename = "")
}
\arguments{
\item{xml}{The XML}
}
\value{
an info list
}
\description{
Get article info from JATS NLM-DTD type XML
}
\keyword{internal}
