% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_apc_fetch.R
\name{open_apc_get_}
\alias{open_apc_get_}
\title{Get single Open APC dataset using GitHub Data API}
\usage{
open_apc_get_(gh_link = NULL)
}
\arguments{
\item{gh_link}{URL to GitHub blob}
}
\description{
Get single Open APC dataset using GitHub Data API
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_license_ind}()},
\code{\link{cr_md_ind}()},
\code{\link{cr_tdm_df}()},
\code{\link{cr_tdm_ind}()},
\code{\link{get_compliant_cc}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{license_normalise}()},
\code{\link{license_report}()},
\code{\link{license_val}()},
\code{\link{open_apc_get}()},
\code{\link{vor_issue}()}
}
\concept{transform}
