% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{openalex}
\alias{openalex}
\title{Get OpenAlex info}
\usage{
openalex(doi)
}
\arguments{
\item{doi}{the DOI of the paper to get info for}
}
\value{
a list of values
}
\description{
Get OpenAlex info
}
\examples{
doi <- "10.7717/peerj.4375"
\dontrun{
  oa_info <- openalex(doi)
}
}
