% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-osf.R
\name{osf_cache_clear}
\alias{osf_cache_clear}
\title{Clear the session cache of OSF listings}
\usage{
osf_cache_clear()
}
\value{
the number of cached listings removed, invisibly
}
\description{
Clear the session cache of OSF listings
}
