% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_file_data}
\alias{osf_file_data}
\title{Structure OSF File Data}
\usage{
osf_file_data(data)
}
\arguments{
\item{data}{the data object from an OSF API call}
}
\value{
a data frame with a subset of data
}
\description{
Structure OSF File Data
}
\keyword{internal}
