% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_headers}
\alias{osf_headers}
\title{OSF Headers}
\usage{
osf_headers(req)
}
\arguments{
\item{req}{an httr2 request object}
}
\value{
the modified request
}
\description{
Adds OSF auth and accept headers to an httr2 request.
}
\keyword{internal}
