% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_node_data}
\alias{osf_node_data}
\title{Structure OSF Node Data}
\usage{
osf_node_data(data)
}
\arguments{
\item{data}{the data object from an OSF API call}
}
\value{
a data frame with a subset of data
}
\description{
Structure OSF Node Data
}
\keyword{internal}
