% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_parse_response}
\alias{osf_parse_response}
\title{Parse an OSF API response into a data frame}
\usage{
osf_parse_response(resp, id, req_type = "guids", pb = NULL)
}
\arguments{
\item{resp}{an httr2 response}

\item{id}{the OSF ID that was requested}

\item{req_type}{the endpoint type used ("guids" or "files")}

\item{pb}{a progress bar}
}
\value{
a single-row data frame
}
\description{
Parse an OSF API response into a data frame
}
\keyword{internal}
