% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_preprint_data}
\alias{osf_preprint_data}
\title{Structure OSF Preprint Data}
\usage{
osf_preprint_data(data)
}
\arguments{
\item{data}{the data object from an OSF API call}
}
\value{
a data frame with a subset of data
}
\description{
Structure OSF Preprint Data
}
\keyword{internal}
