% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_preprint_list}
\alias{osf_preprint_list}
\title{Get A list of preprints from the OSF}
\usage{
osf_preprint_list(
  provider = NULL,
  date_created = NULL,
  date_modified = NULL,
  max_n = 10
)
}
\arguments{
\item{provider}{a vector of the preprint providers, e.g. psyarxiv, socarxiv, edarxiv (see \url{https://osf.io/preprints/discover})}

\item{date_created}{a single date or a vector of two date (min and max)}

\item{date_modified}{a single date or a vector of two date (min and max)}

\item{max_n}{the maximum number of entries to return (will be rounded up to the nearest 10)}
}
\value{
a table of preprint info
}
\description{
Get A list of preprints from the OSF
}
\examples{
\donttest{
  dc <- c("2025-09-01", "2025-10-01")
  pp <- osf_preprint_list("psyarxiv", date_created = dc)
  files <- pp$primary_file
}
}
