% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_reg_data}
\alias{osf_reg_data}
\title{Structure OSF Registration Data}
\usage{
osf_reg_data(data)
}
\arguments{
\item{data}{the data object from an OSF API call}
}
\value{
a data frame with a subset of data
}
\description{
Structure OSF Registration Data
}
\keyword{internal}
