% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper}
\alias{paper}
\title{Create a paper object}
\usage{
paper(id = NULL, ...)
}
\arguments{
\item{id}{The ID of the study}

\item{...}{further arguments to add}
}
\value{
An object with class scivrs_paper
}
\description{
Create a new paper object or load a paper from PDF or XML
}
\keyword{internal}
