% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper_id}
\alias{paper_id}
\title{Get Paper IDs}
\usage{
paper_id(paper)
}
\arguments{
\item{paper}{a paper or paperlist}
}
\value{
a vector of paper_ids
}
\description{
Get Paper IDs
}
\examples{
\dontrun{
psychsci <- papers_load("psychsci", cache = TRUE)
paper_id(psychsci)
}
}
