% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper_schema}
\alias{paper_schema}
\title{Get paper schema}
\usage{
paper_schema()
}
\value{
The schema as a list
}
\description{
Read in the JSON schema for bibr formatted paper objects
}
\examples{
schema <- paper_schema()
schema$`$defs`$info$required
}
