% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper_table}
\alias{paper_table}
\title{Paper tables}
\usage{
paper_table(paper, table, cols = NULL)
}
\arguments{
\item{paper}{a paper or paperlist}

\item{table}{a table name}

\item{cols}{the columns to return from the table (default all columns)}
}
\value{
a merged table
}
\description{
Return a table from a paper object or concatenate tables across a list of paper objects.
}
\examples{
\dontrun{
psychsci <- papers_load("psychsci", cache = TRUE)
biblio <- paper_table(psychsci[1:10], "bib")
xrefs <- paper_table(psychsci[1:10], "xref")
}
}
