% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper_validate}
\alias{paper_validate}
\title{Validate a Paper Object}
\usage{
paper_validate(paper)
}
\arguments{
\item{paper}{a paper object}
}
\value{
TRUE or error
}
\description{
Checks if a paper object conforms to the JSON schema.
}
\examples{
paper <- list(paper_id = "Not a paper object")
tryCatch(
  paper_validate(paper),
  error = \(e) print(e$message)
)

paper <- demopaper()
paper_validate(paper)
}
