% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-papers.R
\name{papers_remove}
\alias{papers_remove}
\title{Remove a cached paper corpus}
\usage{
papers_remove(name)
}
\arguments{
\item{name}{name of the corpus (see \code{\link[=papers_available]{papers_available()}})}
}
\value{
\code{TRUE} invisibly if deleted, \code{FALSE} if not cached
}
\description{
Deletes the locally cached RDS file for a corpus to free disk space.
}
\examples{
\dontrun{
papers_remove("collabra")
}
}
