% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_sanitize}
\alias{path_sanitize}
\title{Sanitize File Path}
\usage{
path_sanitize(
  path,
  replacement = "_",
  remove_whitespace = TRUE,
  keep_sep = TRUE
)
}
\arguments{
\item{path}{the path to sanitize (can be a vector of paths)}

\item{replacement}{the character to replace invalid characters with}

\item{remove_whitespace}{whether to include whitespace as a problem}

\item{keep_sep}{whether to keep the path separator /}
}
\value{
the sanitized vector
}
\description{
Make sure user-input file names are not problematic.
}
\examples{
path <- "/My Files/x><y.pdf"
path_sanitize(path)
path_sanitize(path, replacement = "~")
path_sanitize(path, remove_whitespace = FALSE)
path_sanitize(path, keep_sep = FALSE)
}
