% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretest.R
\name{pretest}
\alias{pretest}
\alias{tabulate_metacheckable}
\alias{is_metacheckable}
\alias{assert_metacheckable}
\alias{report_metacheckable}
\alias{is_in_limit}
\alias{is_doi_cr_md}
\alias{is_doi_cr_type}
\title{Pretest acceptable DOIs}
\usage{
tabulate_metacheckable(x, ...)

is_metacheckable(x, ...)

assert_metacheckable(x, ...)

report_metacheckable(x, lang = mc_langs, ...)

is_in_limit(x, limit = 1000L)

is_doi_cr_md(x)

is_doi_cr_type(x, type = types_allowed)
}
\arguments{
\item{x}{A vector created by, or convertable to \code{\link[biblids:doi]{doi()}}.}

\item{...}{Additional arguments passed to the predicate functions.}

\item{lang}{Character scalar giving the anguage of the report.
\emph{Not} reactive as in \code{\link[=mcControlsServer]{mcControlsServer()}}.}

\item{limit}{Scalar integer, giving the number of DOIs to be submitted to the metacheck.}

\item{type}{Character scalar, must be one of the types from \code{\link[rcrossref:cr_types]{rcrossref::cr_types()}}.}
}
\value{
A tibble with one column for each of the predicate results listed above.
\code{TRUE} always means that a DOI passes a criterion.

\code{NA} can indicate that the test:
\itemize{
\item was not applicable, because a previous predicate was \code{FALSE}
}
}
\description{
Pretest acceptable DOIs
}
\details{
Before even checking what metadata has been deposited for a DOI, a DOI has to pass a number of criteria to be eligible for metacheck.

Eligible DOIs must be:
\enumerate{
\item non-missing (\code{not_na}).
For details, see \code{biblids::doi()}.
\item non-duplicated (\code{unique}).
Every 2nd and later repetition of a DOI fails this test.
For details, see \code{biblids::doi()}.
\item within the limit of 1000 DOIs (\code{within_limits}).
This limit is set for the web application to avoid misuse.
Please contact \href{https://subugoe.github.io/metacheck/articles/help.html}{support} if you need to test more DOIs.
\item resolvable on DOI.org (\code{doi_org_found}).
For details, see \code{biblids::is_doi_found()}.
\item have been deposited by the Crossref registration agency (as per doi.org) (\code{from_cr}).
DOIs registered by an agency other than Crossref will not have metadata on Crossref.
\item have metadata deposited on Crossref (\code{cr_md}).
\item have to resolve to a journal article (\code{article}).
}

These criteria are tested in this order, and assumed to be transitive but asymmetric in the order given.
For example, if a DOI is not found on DOI.org, it is assumed that it also \emph{cannot} be known to Crossref.

This must necessarily be the case for, say,
syntactically invalud DOIs (which cannot have metadata),
but is more of an assumption when it comes to the consistency between
different data sources.
For example, it is conceivable, though implausible,
that a DOI might not resolve on DOI.org,
but is still listed on Crossref.
}
\section{Functions}{
\itemize{
\item \code{tabulate_metacheckable}: Tabulate results

\item \code{is_metacheckable}: Are all predicates \code{TRUE}?

\item \code{assert_metacheckable}: Assert metacheckability

\item \code{report_metacheckable}: Report summary in markdown

\item \code{is_in_limit}: Is the DOI in the limit?

\item \code{is_doi_cr_md}: Is there metadata for the DOI on Crossref?

\item \code{is_doi_cr_type}: Is the DOI a type (per Crossref)?
}}

\seealso{
Other import: 
\code{\link{cr_works2}()},
\code{\link{get_cr_md}()}
}
\concept{import}
