% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretest.R
\name{pretests}
\alias{pretests}
\alias{tabulate_metacheckable}
\alias{is_metacheckable}
\alias{assert_metacheckable}
\alias{report_metacheckable}
\title{Pretest acceptable DOIs}
\usage{
pretests(...)

tabulate_metacheckable(x, ...)

is_metacheckable(x, ...)

assert_metacheckable(x, ...)

report_metacheckable(x, lang = mc_langs, ...)
}
\arguments{
\item{...}{Additional arguments passed to the predicate functions.}

\item{x}{A vector created by, or convertable to \code{\link[biblids:doi]{doi()}}.}

\item{lang}{Character scalar giving the anguage of the report.
\emph{Not} reactive as in \code{\link[=mcControlsServer]{mcControlsServer()}}.}
}
\value{
\code{TRUE} indicates that the test passes the criterion, \code{FALSE} otherwise.
\code{NA} can indicates that the test was not applicable,
because a previous predicate was \code{FALSE}.
}
\description{
DOIs have to meet some pretests to have a meaningful metadata compliance.
}
\details{
These criteria are tested,
and assumed to be transitive but asymmetric in the order given.
For example, if a DOI is not found on DOI.org,
it is assumed that it also \emph{cannot} be known to Crossref.

This must necessarily be the case for, say,
syntactically invalid DOIs (which cannot have metadata),
but is more of an assumption when it comes to the consistency between
different data sources.
For example, it is conceivable, though implausible,
that a DOI might not resolve on DOI.org,
but is still listed on Crossref.

For details about DOI predicates (\code{\link[=is.na]{is.na()}} etc.), see \code{\link[biblids:doi]{biblids::doi()}}.
For details about custom metacheck predicates,
see \code{\link[=mc_predicates]{mc_predicates()}}.
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{tabulate_metacheckable}: Tabulate results
}
\itemize{
\item \code{is_metacheckable}: Are all predicates \code{TRUE}?
}
\itemize{
\item \code{assert_metacheckable}: Assert metacheckability
}
\itemize{
\item \code{report_metacheckable}: Report summary in markdown
}}}

\examples{
pretests()
}
\seealso{
Other import: 
\code{\link{cr_works2}()},
\code{\link{get_cr_md}()}
}
\concept{import}
