% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{print.metacheck_module_output}
\alias{print.metacheck_module_output}
\title{Print Module Output}
\usage{
\method{print}{metacheck_module_output}(x, ...)
}
\arguments{
\item{x}{The metacheck_module_output object}

\item{...}{Additional parameters for \code{module_report()}}
}
\description{
Print Module Output
}
\keyword{internal}
