% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{print.metacheck_validate}
\alias{print.metacheck_validate}
\title{Print Validation List Object}
\usage{
\method{print}{metacheck_validate}(x, ...)
}
\arguments{
\item{x}{The metacheck_validate object}

\item{...}{Additional parameters for print}
}
\description{
Print Validation List Object
}
\keyword{internal}
