% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.scivrs_paperlist}
\alias{print.scivrs_paperlist}
\title{Print PaperList Object}
\usage{
\method{print}{scivrs_paperlist}(x, ...)
}
\arguments{
\item{x}{The scivrs_paperlist object}

\item{...}{Additional parameters for print}
}
\description{
Print PaperList Object
}
\keyword{internal}
