% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{process_full_text}
\alias{process_full_text}
\title{Process full text table}
\usage{
process_full_text(full_text)
}
\arguments{
\item{full_text}{a table of the full text}
}
\value{
a data frame
}
\description{
Process full text table
}
\keyword{internal}
