% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-psycharchives.R
\name{psycharchives_links}
\alias{psycharchives_links}
\title{Find PsychArchives Links in Papers}
\usage{
psycharchives_links(paper)
}
\arguments{
\item{paper}{a paper object or paperlist object}
}
\value{
a table with the PsychArchives url in the first (href) column
}
\description{
Find PsychArchives Links in Papers
}
\examples{
\dontrun{
psycharchives_links(psychsci)
}
}
