% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_ind.R
\name{react_ind_table}
\alias{react_ind_table}
\title{Reactable represenation of metadata indicators}
\usage{
react_ind_table(ind_table, fill_col = "#fc5185")
}
\arguments{
\item{ind_table}{tibble compliance overview table}

\item{fill_col}{fill color for bar charts (hex code)}
}
\description{
Inspired from \url{https://glin.github.io/reactable/articles/building-twitter-followers.html}
}
