% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibr.R
\name{read}
\alias{read}
\title{Read in grobid XML or bibr JSON}
\usage{
read(file_path)
}
\arguments{
\item{file_path}{path to a directory containing XML and/or JSON files, or a vector of paths}
}
\value{
a paper or paperlist
}
\description{
Read in grobid XML or bibr JSON
}
