% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cermine.R
\name{read_cermine}
\alias{read_cermine}
\title{Get paper from cermine XML file}
\usage{
read_cermine(filename)
}
\arguments{
\item{filename}{the path to the XML file, a vector of file paths, or the path to a directory containing XML files}
}
\value{
A paper object with class scivrs_paper, or a list of paper objects
}
\description{
Convert PDFs to cermine XML at \url{http://cermine.ceon.pl}
}
\examples{
# paper <- read_cermine(filename)
}
