% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cermine.R
\name{read_cermine_xml}
\alias{read_cermine_xml}
\title{Read in cermine XML}
\usage{
read_cermine_xml(filename)
}
\arguments{
\item{filename}{The path to the XML file to be read}
}
\value{
An XML object
}
\description{
Read in cermine XML
}
\keyword{internal}
