% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{read_file}
\alias{read_file}
\title{Read file of unknown type}
\usage{
read_file(path)
}
\arguments{
\item{path}{the path to the file}
}
\value{
a data frame
}
\description{
Read file of unknown type
}
\keyword{internal}
