% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_text.R
\name{read_text}
\alias{read_text}
\title{Get paper from text}
\usage{
read_text(txt, id = "")
}
\arguments{
\item{txt}{The text to read into a paper object, or a path to a text file}

\item{id}{The ID to use for the paper (e.g., a DOI) defaults to the filename if txt is a path}
}
\value{
A paper object with class scivrs_paper
}
\description{
Get paper from text
}
\examples{
paper <- read_text("This is my paper. That's it!")
}
