% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacheck-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pb}
\alias{verbose}
\alias{\%||\%}
\alias{message}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{svutils}{\code{\link[svutils:op-null-default]{\%||\%}}, \code{\link[svutils]{message}}, \code{\link[svutils]{pb}}, \code{\link[svutils]{verbose}}}
}}

