% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcheck-local.R
\name{regcheck_stop_local}
\alias{regcheck_stop_local}
\title{Stop the local RegCheck server}
\usage{
regcheck_stop_local()
}
\value{
invisibly NULL
}
\description{
Kills the background process started by \code{\link[=regcheck_start_local]{regcheck_start_local()}}.
}
