% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{render_email}
\alias{render_email}
\title{Send out email}
\usage{
render_email(
  dois = readLines(system.file("data-raw/tu_dois.txt", package = "metacheck"))[1:10]
)
}
\arguments{
\item{dois}{Character vector of DOIs}
}
\description{
Send out email
}
