% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{report_app}
\alias{report_app}
\title{Launch Report App}
\usage{
report_app(quiet = FALSE, ...)
}
\arguments{
\item{quiet}{whether to show debugging messages in the console}

\item{...}{arguments to pass to shiny::runApp}
}
\value{
NULL (invisibly)
}
\description{
Launch the Report app: upload a PDF and generate a report with one
click, with privacy options for what is sent to external servers.
}
\examples{
\dontrun{
report_app()
}

}
