% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_harsh.R
\name{report_harsh}
\alias{report_harsh}
\title{Create a Report}
\usage{
report_harsh(
  paper,
  modules = c("power_harsh", "repo_check_harsh", "code_check_harsh",
    "stat_p_exact_harsh", "stat_p_nonsig_harsh", "stat_effect_size_harsh",
    "marginal_harsh", "ref_accuracy_harsh"),
  output_file = paste0(paper$paper_id, "_report.", output_format),
  output_format = c("html", "qmd"),
  args = list()
)
}
\arguments{
\item{paper}{a paper object or a paperlist object}

\item{modules}{a vector of modules to run (names for built-in modules or paths for custom modules)}

\item{output_file}{the name of the output file}

\item{output_format}{the format to create the report in}

\item{args}{a list of arguments to pass to modules (see Details)}
}
\value{
the file path the report is saved to
}
\description{
Run specified modules on a paper and generate a report in quarto (qmd), html, or pdf format.
}
\details{
Pass arguments to modules in a named list of lists, using the same names as the \code{modules} argument. You only need to specify modules with arguments.

\if{html}{\out{<div class="sourceCode">}}\preformatted{args <- list(power = list(seed = 8675309))
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
paper <- demopaper()
report(paper)
}
}
