% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R, R/report_harsh.R, R/report_kind.R
\name{report_module_run}
\alias{report_module_run}
\title{Run modules for a report}
\usage{
report_module_run(paper, modules, args = list())

report_module_run(paper, modules, args = list())

report_module_run(paper, modules, args = list())
}
\arguments{
\item{paper}{a paper object}

\item{modules}{a vector of modules to run}

\item{args}{optional list of arguments to pass to modules}
}
\value{
a list of module outputs

a list of module outputs

a list of module outputs
}
\description{
Runs modules in order on the paper and orders by section and traffic light.

Runs modules in order on the paper and orders by section and traffic light.

Runs modules in order on the paper and orders by section and traffic light.
}
\details{
Pass arguments to modules in a named list of lists, using the same names as the \code{modules} argument. You only need to specify modules with arguments.

\if{html}{\out{<div class="sourceCode">}}\preformatted{args <- list(power = list(seed = 8675309))
}\if{html}{\out{</div>}}

Pass arguments to modules in a named list of lists, using the same names as the \code{modules} argument. You only need to specify modules with arguments.

\if{html}{\out{<div class="sourceCode">}}\preformatted{args <- list(power = list(seed = 8675309))
}\if{html}{\out{</div>}}

Pass arguments to modules in a named list of lists, using the same names as the \code{modules} argument. You only need to specify modules with arguments.

\if{html}{\out{<div class="sourceCode">}}\preformatted{args <- list(power = list(seed = 8675309))
}\if{html}{\out{</div>}}
}
\examples{
paper <- demopaper()
modules <- c("stat_p_exact", "stat_p_nonsig")
module_output <- report_module_run(paper, modules)
paper <- demopaper()
modules <- c("stat_p_exact", "stat_p_nonsig")
module_output <- report_module_run(paper, modules)
paper <- demopaper()
modules <- c("stat_p_exact", "stat_p_nonsig")
module_output <- report_module_run(paper, modules)
}
