% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_kind.R
\name{report_qmd_kind}
\alias{report_qmd_kind}
\title{Create Report from Module Output}
\usage{
report_qmd_kind(module_output, paper = list())
}
\arguments{
\item{module_output}{a list of module output (usually from \code{report_module_run()})}

\item{paper}{a paper object}
}
\value{
report text
}
\description{
Create Report from Module Output
}
