% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchbox.R
\name{researchbox_info}
\alias{researchbox_info}
\title{Retrieve info from ResearchBox by URL}
\usage{
researchbox_info(rb_url)
}
\arguments{
\item{rb_url}{a ResearchBox URL}
}
\value{
a data frame of information
}
\description{
Retrieve info from ResearchBox by URL
}
\keyword{internal}
