% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchbox.R
\name{researchbox_links}
\alias{researchbox_links}
\title{Find ResearchBox Links in Papers}
\usage{
researchbox_links(paper)
}
\arguments{
\item{paper}{a paper object or paperlist object}
}
\value{
a table with the ResearchBox url in the first (text) column
}
\description{
Find ResearchBox Links in Papers
}
\examples{
researchbox_links(psychsci)
}
