% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchbox.R
\name{researchbox_retrieve}
\alias{researchbox_retrieve}
\title{Retrieve info from ResearchBox by URL}
\usage{
researchbox_retrieve(rb_url, id_col = 1)
}
\arguments{
\item{id_col}{the index or name of the column that contains ResearchBox URLs, if id is a table}

\item{ap_url}{an ResearchBox URL, or a table containing them (e.g., as created by \code{researchbox_links()})}
}
\value{
a data frame of information
}
\description{
Retrieve info from ResearchBox by URL
}
\examples{
\donttest{
  # get info on one OSF node
  researchbox_retrieve("https://researchbox.org/801")
}
}
