% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communicate.R
\name{runMetacheck}
\alias{runMetacheck}
\title{Start web application}
\usage{
runMetacheck(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[shiny:runApp]{shiny::runApp}}
  \describe{
    \item{\code{appDir}}{The application to run. Should be one of the following:
\itemize{
\item A directory containing \code{server.R}, plus, either \code{ui.R} or
a \code{www} directory that contains the file \code{index.html}.
\item A directory containing \code{app.R}.
\item An \code{.R} file containing a Shiny application, ending with an
expression that produces a Shiny app object.
\item A list with \code{ui} and \code{server} components.
\item A Shiny app object created by \code{\link[shiny:shinyApp]{shinyApp()}}.
}}
    \item{\code{port}}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}
    \item{\code{launch.browser}}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only. This value of this parameter can also be a
function to call with the application's URL.}
    \item{\code{host}}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}
    \item{\code{workerId}}{Can generally be ignored. Exists to help some editions of
Shiny Server Pro route requests to the correct process.}
    \item{\code{quiet}}{Should Shiny status messages be shown? Defaults to FALSE.}
    \item{\code{display.mode}}{The mode in which to display the application. If set to
the value \code{"showcase"}, shows application code and metadata from a
\code{DESCRIPTION} file in the application directory alongside the
application. If set to \code{"normal"}, displays the application normally.
Defaults to \code{"auto"}, which displays the application in the mode given
in its \code{DESCRIPTION} file, if any.}
    \item{\code{test.mode}}{Should the application be launched in test mode? This is
only used for recording or running automated tests. Defaults to the
\code{shiny.testmode} option, or FALSE if the option is not set.}
  }}
}
\description{
Start web application
}
\seealso{
Other communicate: 
\code{\link{emailReport}()},
\code{\link{email}},
\code{\link{mcApp}()},
\code{\link{mcControls}},
\code{\link{report}}
}
\concept{communicate}
