% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractionwatch.R
\name{rw_update}
\alias{rw_update}
\title{Update retractionwatch}
\usage{
rw_update()
}
\value{
the path to the data frame (invisibly)
}
\description{
metacheck comes with a built-in data frame called \code{retractionwatch}. We update it regularly, but you can use this function to download the newest version. The download is >50MB, but this function will summarise the information into a smaller version (~0.5 MB) and delete the original file.
}
