% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{send_email}
\alias{send_email}
\title{Send out email}
\usage{
send_email(to, email)
}
\arguments{
\item{to}{A vector of email addresses serving as primary recipients for the
message. For secondary recipients, use the \code{cc} and \code{bcc} arguments. A
named character vector can be used to specify the recipient names along
with the their email address (e.g., \code{c("Jane Doe" = "jane_doe@example.com")}).}

\item{email}{The email message object, as created by the \code{\link[blastula:compose_email]{compose_email()}}
function. The object's class is \code{email_message}.}
}
\description{
Send out email
}
