% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{set_llm_max_calls}
\alias{set_llm_max_calls}
\title{Set the maximum number of calls to the LLM}
\usage{
set_llm_max_calls(n = 10)
}
\arguments{
\item{n}{The maximum number of calls that the llm() function can make}
}
\description{
Set the maximum number of calls to the LLM
}
