% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scienceverse.R
\name{study}
\alias{study}
\title{Create a study object}
\usage{
study(name = "Demo Study", ...)
}
\arguments{
\item{name}{The name of the study}

\item{...}{further arguments to add}
}
\value{
A study object with class scivrs_study
}
\description{
Create a new study object
}
\keyword{internal}
