% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.scivrs_paperlist}
\alias{[.scivrs_paperlist}
\title{Subset PaperList Object}
\usage{
\method{[}{scivrs_paperlist}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{The scivrs_paperlist object}

\item{...}{Additional parameters for print}

\item{drop}{relevant for matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples).}
}
\description{
Subset PaperList Object
}
\keyword{internal}
