% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{tabulate_metacheckable}
\alias{tabulate_metacheckable}
\alias{is_metacheckable}
\alias{assert_metacheckable}
\title{Report and filter acceptable DOIs}
\usage{
tabulate_metacheckable(x, ...)

is_metacheckable(x, ...)

assert_metacheckable(x, ...)
}
\arguments{
\item{x}{A vector created by, or convertable to \code{\link[biblids:doi]{doi()}}.}

\item{...}{Additional arguments passed to predicate functions.}
}
\value{
A tibble with one column for each of predicate results.
\code{TRUE} always means that a DOI passes a criterion.

Tested in order:
\enumerate{
\item \code{not_na} following \code{\link[biblids:doi]{biblids::doi()}}.
\item \code{unique}
\code{FALSE} for every 2nd and later repetition of a DOI.
DOIs are compared using \code{\link[biblids:doi]{biblids::doi()}} logic.
\item \code{within_limits}
whether remaining DOIs are within the package limit for requests.
\item \code{doi_org_found}
whether remaining DOIs are found on DOI.org,
using \code{\link[biblids:doi_api]{biblids::is_doi_found()}}.
\item \code{resolvable}
whether remaining DOIs are resolvable on DOI.org,
using \code{\link[biblids:doi_api]{biblids::is_doi_resolvable()}}.
\item \code{from_cr}
whether remaining DOIs have been deposited by the Crossref
registration agency.
}

\code{NA} can indicate that the test:
\itemize{
\item was not applicable, because a previous predicate was \code{FALSE}
}
}
\description{
Report and filter acceptable DOIs
}
\details{
Before even checking what metadata has been deposited for a DOI,
a DOI has to pass a number of tests to be eligible for metacheck.

The below criteria are assumed to be transitive but asymmetric
in the order given.
For example, if a DOI is not found on DOI.org,
it is assumed that it is also not known to crossref and all other tests
are also assumed to be \code{NA}.
This must necessarily be the case for, say, syntactically invalid DOIs,
but is more of an assumption when it comes to the consistency between
different data sources.
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{is_metacheckable}: Are all checks \code{TRUE}?
}
\itemize{
\item \code{assert_metacheckable}: Assert metacheckable
}}}

\seealso{
Other import: 
\code{\link{get_cr_md}()}
}
\concept{import}
