% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_tdm.R
\name{tdm_metrics}
\alias{tdm_metrics}
\title{TDM MIME-Type metrics}
\usage{
tdm_metrics(.md = NULL, .gt = TRUE, .color = "#9B0056")
}
\arguments{
\item{.md}{list, returned from \code{\link[=cr_compliance_overview]{cr_compliance_overview()}}}

\item{.gt}{logical, should results be represented as styled html tables or
as tibble. Default .gt = TRUE.}

\item{.color}{character, hex color code for styling HTML proportional bar chart}
}
\description{
Presents MIME-Types for full-text versions of version-of-records.
}
\details{
Publishers can share more than one TDM-link per full-text using
Crossref metadata. This function applies full counting, where
full-texts were assigned to more than one MIME type were counted once
for each. Relative number is calculated by the number of checked articles.
Articles without compliant TDM information are tagged as
"No TDM links for version-of-records".
}
\examples{
\dontrun{
## Some DOIs from OA articles
my_dois <- c("10.5194/wes-2019-70", "10.1038/s41598-020-57429-5",
  "10.3389/fmech.2019.00073", "10.1038/s41598-020-62245-y",
  "10.1109/JLT.2019.2961931")

# Workflow:
# First, obtain metadata from Crossref API
req <- get_cr_md(test_dois)

# Then, check article-level compliance

# Finally, obtain TDM metrics overview
tdm_metrics(out)
# change bar chart color
tdm_metrics(out, .color = "green")
# as tibble
tdm_metrics(out, .gt = FALSE)
}
}
