% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{tei_authors}
\alias{tei_authors}
\title{Get author info from TEI type XML}
\usage{
tei_authors(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
an author list
}
\description{
Get author info from TEI type XML
}
\keyword{internal}
