% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{tei_bib}
\alias{tei_bib}
\title{Get bibliography from TEI type XML}
\usage{
tei_bib(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
bib table
}
\description{
Get bibliography from TEI type XML
}
\keyword{internal}
