% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{tei_info}
\alias{tei_info}
\title{Get article info from TEI type XML}
\usage{
tei_info(xml, filename = "")
}
\arguments{
\item{xml}{The XML}
}
\value{
an info list
}
\description{
Get article info from TEI type XML
}
\keyword{internal}
