% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tei.R
\name{tei_to_bibr}
\alias{tei_to_bibr}
\title{Convert Grobid TEI XML file to bibr format}
\usage{
tei_to_bibr(xml_file)
}
\arguments{
\item{xml_file}{the XML file}
}
\value{
a paper object
}
\description{
Convert Grobid TEI XML file to bibr format
}
