% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{tei_xrefs}
\alias{tei_xrefs}
\title{Get cross references from TEI type XML}
\usage{
tei_xrefs(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
xrefs table
}
\description{
Get cross references from TEI type XML
}
\keyword{internal}
