% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{text_authors}
\alias{text_authors}
\title{Get author info from plain text}
\usage{
text_authors(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
an author list
}
\description{
Get author info from plain text
}
\keyword{internal}
