% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{text_bib}
\alias{text_bib}
\title{Get bibliography from text file}
\usage{
text_bib(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
bib table
}
\description{
Get bibliography from text file
}
\keyword{internal}
