% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_expand.R
\name{text_expand}
\alias{text_expand}
\alias{expand_text}
\title{Expand text}
\usage{
text_expand(
  results_table,
  paper,
  expand_to = c("sentence", "paragraph", "div", "section"),
  plus = 0,
  minus = 0
)

expand_text(
  results_table,
  paper,
  expand_to = c("sentence", "paragraph", "div", "section"),
  plus = 0,
  minus = 0
)
}
\arguments{
\item{results_table}{the table to expand}

\item{paper}{a metacheck paper object or a list of paper objects to look up the expanded text from}

\item{expand_to}{whether to expand to the sentence, paragraph, div, or section level}

\item{plus}{append additional sentences after the target expansion}

\item{minus}{prepend additional sentences before the target expansion}
}
\value{
a results table with the expanded text
}
\description{
If you have a table resulting from \code{text_search()} or a module return object, you can expand the text column to the full sentence, paragraph, or section. You can also set \code{plus} and \code{minus} to append and prepend sentences to the result (only when \code{expand_to} is "sentence").
}
\examples{
# single paper search
paper <- demopaper()
res_tbl <- text_search(paper, "p =", return = "match")
expanded <- text_expand(res_tbl, paper)

# multiple paper search
\dontrun{
papers <- papers_load("psychsci", cache = TRUE)
res_tbl <- text_search(papers, "replicate")
expanded <- text_expand(res_tbl, papers, plus = 1, minus = 1)
}
}
