% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{text_full_text}
\alias{text_full_text}
\title{Get full text from plain text}
\usage{
text_full_text(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
a data frame with all text
}
\description{
Get full text from plain text
}
\keyword{internal}
