% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{text_info}
\alias{text_info}
\title{Get article info from plain text}
\usage{
text_info(xml, filename = "")
}
\arguments{
\item{xml}{The XML}
}
\value{
an info list
}
\description{
Get article info from plain text
}
\keyword{internal}
