% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{text_xrefs}
\alias{text_xrefs}
\title{Get cross references from text file}
\usage{
text_xrefs(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
refs table
}
\description{
Get cross references from text file
}
\keyword{internal}
