% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{tl_accuracy}
\alias{tl_accuracy}
\title{Traffic Light Accuracy}
\usage{
tl_accuracy(validation, yes = "green", no = "red")
}
\arguments{
\item{validation}{the validation object}

\item{yes}{the value of traffic lights for detection (usually "green")}

\item{no}{the value of traffic lights for no detection (usually "red")}
}
\value{
a list of accuracy parameters
}
\description{
Traffic Light Accuracy
}
