% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_cc.R
\name{vor_issue}
\alias{vor_issue}
\title{Extract records with CC license not applied to version of records}
\usage{
vor_issue(license_df, compliant_dois)
}
\arguments{
\item{license_df}{normalized license metadata from \code{\link[=license_check]{license_check()}}}

\item{compliant_dois}{DOIs representing records with valid CC info}
}
\description{
In order to be compliant, CC license has to apply to version of record and must be valid without delay.
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_license_ind}()},
\code{\link{cr_md_ind}()},
\code{\link{cr_tdm_df}()},
\code{\link{cr_tdm_ind}()},
\code{\link{get_compliant_cc}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{license_normalise}()},
\code{\link{license_report}()},
\code{\link{license_val}()},
\code{\link{open_apc_get_}()},
\code{\link{open_apc_get}()}
}
\concept{transform}
