% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{word_authors}
\alias{word_authors}
\title{Get author info from  Word}
\usage{
word_authors(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
an author list
}
\description{
Get author info from  Word
}
\keyword{internal}
