% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{word_bib}
\alias{word_bib}
\title{Get references/citations from Word .docx}
\usage{
word_bib(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
an list of two tables: references and citations
}
\description{
Get references/citations from Word .docx
}
\keyword{internal}
