% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{word_full_text}
\alias{word_full_text}
\title{Get full text from  Word}
\usage{
word_full_text(filename)
}
\arguments{
\item{filename}{The filename}
}
\value{
a data frame with all text
}
\description{
Get full text from  Word
}
\keyword{internal}
