% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{word_info}
\alias{word_info}
\title{Get article info from Word}
\usage{
word_info(xml, filename = "")
}
\arguments{
\item{xml}{The XML}
}
\value{
an info list
}
\description{
Get article info from Word
}
\keyword{internal}
