% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{word_xrefs}
\alias{word_xrefs}
\title{Get cross references from Word .docx}
\usage{
word_xrefs(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
xrefs table
}
\description{
Get cross references from Word .docx
}
\keyword{internal}
