% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{xlsx_path}
\alias{xlsx_path}
\title{Temp path to write xlsx to}
\usage{
xlsx_path(session_id = NULL)
}
\arguments{
\item{session_id}{Character vector to identify current shiny session}
}
\description{
Temp path to write xlsx to
}
\keyword{internal}
