% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{xml2bib}
\alias{xml2bib}
\title{Parse XML bib format to bibtex}
\usage{
xml2bib(ref)
}
\arguments{
\item{ref}{the biblStruct xml object}
}
\value{
a bibentry
}
\description{
Parse XML bib format to bibtex
}
\keyword{internal}
