% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{xml_date}
\alias{xml_date}
\title{Find and return date info from XML}
\usage{
xml_date(xml, xpath = ".//string-date")
}
\arguments{
\item{xml}{the xml node}

\item{xpath}{a string containing an xpath expression}
}
\value{
text
}
\description{
Find and return date info from XML
}
\keyword{internal}
