% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-zenodo.R
\name{zenodo_info}
\alias{zenodo_info}
\title{Retrieve info from Zenodo by URL}
\usage{
zenodo_info(zenodo_url, id_col = 1, pb = NULL, cache = FALSE)
}
\arguments{
\item{zenodo_url}{an Zenodo URL, or a table containing them (e.g., as created by \code{zenodo_links()})}

\item{id_col}{the index or name of the column that contains Zenodo URLs, if id is a table}

\item{pb}{a progress bar passed from another function}

\item{cache}{if \code{TRUE}, reuse a previously cached listing for a record
already looked up (see \code{\link[=repo_info_cache]{repo_info_cache()}}) instead of re-querying
Zenodo's API. Off by default.}
}
\value{
a data frame of information
}
\description{
Retrieve info from Zenodo by URL
}
\examples{
\dontrun{
  # get info on one zenodo link
  zenodo_info("https://doi.org/10.5281/zenodo.18648142")
}
}
