% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibr.R
\name{bibr_convert}
\alias{bibr_convert}
\title{Process a paper using the bibr API}
\usage{
bibr_convert(
  file_path,
  save_dir = ".",
  api_url = "https://api.bibr.metacheck.app",
  api_key = Sys.getenv("BIBR_API"),
  start_page = 1,
  end_page = Inf
)
}
\arguments{
\item{file_path}{Path to the document file, or a directory of documents}

\item{save_dir}{Path to a directory in which to save the JSON file}

\item{api_url}{Base URL of the API}

\item{api_key}{Key to access bibr}

\item{start_page}{First page of the file to extract}

\item{end_page}{Last page of the file to extract}
}
\value{
A list of parsed information
}
\description{
Process a paper using the bibr API
}
\keyword{internal}
