% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_cc.R
\name{cc_compliance_metrics}
\alias{cc_compliance_metrics}
\title{CC compliance metrics overview}
\usage{
cc_compliance_metrics(cc_license_check = NULL)
}
\arguments{
\item{cc_license_check}{tibble, result from \code{\link[=license_check]{license_check()}}}
}
\description{
Presents the result of the CC compliance check (absolute and relative)
}
\seealso{
\code{\link[=license_check]{license_check()}}
}
