% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_cc.R
\name{cc_metrics}
\alias{cc_metrics}
\title{CC license variants metrics overview}
\usage{
cc_metrics(.md = NULL, .gt = TRUE, .color = "#F3A9BB")
}
\arguments{
\item{.md}{list, returned from \code{\link[=cr_compliance_overview]{cr_compliance_overview()}}}

\item{.gt}{logical, should results be represented as styled html tables or
as tibble. Default .gt = TRUE.}

\item{.color}{character, hex color code for styling HTML proportional bar chart}
}
\description{
Presents normalised CC licence variants like BY (absolute and relative)
}
