% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_cc.R
\name{cc_metrics}
\alias{cc_metrics}
\title{CC license variants metrics overview}
\usage{
cc_metrics(cc_license_check = NULL)
}
\arguments{
\item{cc_license_check}{tibble, result from \code{\link[=license_check]{license_check()}}}
}
\description{
Presents normalised CC licence variants like BY (absolute and relative)
}
\examples{
\dontrun{
my_dois <- c("10.5194/wes-2019-70", "10.1038/s41598-020-57429-5",
  "10.3389/fmech.2019.00073", "10.1038/s41598-020-62245-y",
  "10.1109/JLT.2019.2961931")

# Workflow:
# First, obtain metadata from Crossref API
req <- get_cr_md(my_dois)

# Then, check article-level compliance

# Obtain CC variants metrics
cc_metrics(out$> out$cc_license_check)
}
}
