% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_check.R
\name{code_file_refs}
\alias{code_file_refs}
\title{Get files referenced in code}
\usage{
code_file_refs(code_text, lang = c("R", "SPSS", "SAS", "Stata"))
}
\arguments{
\item{code_text}{the code text for a single file}

\item{lang}{the language (we only currently handle R, SPSS, SAS, Stata)}
}
\value{
a vector of files that are referenced in the code
}
\description{
Get files referenced in code
}
\examples{
code_text <- c(
  'source("functions.R")',
  'a <- "bread"',
  'b <- read.csv("file.csv")'
)
code_file_refs(code_text, "R")

}
