% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_check.R
\name{code_lang}
\alias{code_lang}
\title{Detect Code Language}
\usage{
code_lang(file_name)
}
\arguments{
\item{file_name}{a vector of file names}
}
\value{
a vector of languages
}
\description{
Detects code language used in files, only for languages metacheck currently processes (R, SAS, SPSS, Stata).
}
\examples{
file_name <- "file.R"
code_lang(file_name)

file_name <- c("file.Rmd", "file.SAS", "file.r", "file.qmd", "file.txt")
code_lang(file_name)
}
