% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_check.R
\name{code_line_stats}
\alias{code_line_stats}
\title{Get Code Composition Stats}
\usage{
code_line_stats(
  code_text,
  lang = c("R", "Python", "SPSS", "SAS", "Stata", "Mplus", "MATLAB")
)
}
\arguments{
\item{code_text}{the code text for a single file}

\item{lang}{the language (we only currently handle R, Python, SPSS, SAS,
Stata, Mplus, MATLAB)}
}
\value{
list with items \code{total_lines}, \code{comment_lines}, \code{code_lines},
\code{percent_comment}, and (Python only) \code{has_docstring}
}
\description{
Get Code Composition Stats
}
\examples{
code_text <- c(
  "library(dplyr)",
  "",
  "# this line is a comment",
  "a <- 1"
)
code_line_stats(code_text, "R")
}
