% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_overview.R
\name{cr_compliance_overview}
\alias{cr_compliance_overview}
\title{Get compliance overview}
\usage{
cr_compliance_overview(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Get compliance overview
}
\seealso{
Other transform: 
\code{\link{cc}},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_tdm_df}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{mc_predicates}},
\code{\link{vor_issue}()}
}
\concept{transform}
