% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_funder.R
\name{cr_funder_df}
\alias{cr_funder_df}
\title{Get funder list from Crossref metadata}
\usage{
cr_funder_df(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Get funder list from Crossref metadata
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_license_ind}()},
\code{\link{cr_md_ind}()},
\code{\link{cr_tdm_df}()},
\code{\link{cr_tdm_ind}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{vor_issue}()}
}
\concept{transform}
