% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_overview.R
\name{cr_has_orcid}
\alias{cr_has_orcid}
\title{Check for ORCIDs}
\usage{
cr_has_orcid(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Check for ORCIDs
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_tdm_df}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{vor_issue}()}
}
\concept{transform}
