% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_tdm.R
\name{cr_tdm_df}
\alias{cr_tdm_df}
\title{Get text-mining links from Crossref metadata}
\usage{
cr_tdm_df(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Unnest TDM information for version-of-records from Crossref metadata
}
\seealso{
Other transform: 
\code{\link{cc}},
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{vor_issue}()}
}
\concept{transform}
