% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-dataverse.R
\name{dataverse_info}
\alias{dataverse_info}
\title{Retrieve info from Dataverse by URL}
\usage{
dataverse_info(dataverse_url, id_col = 1, pb = NULL)
}
\arguments{
\item{dataverse_url}{a Dataverse dataset URL (or DOI, if \code{host} is also
given), or a table containing them (e.g., as created by
\code{\link[=dataverse_links]{dataverse_links()}})}

\item{id_col}{the index or name of the column that contains Dataverse URLs,
if \code{dataverse_url} is a table}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from Dataverse by URL
}
\examples{
\dontrun{
  dataverse_info("https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/GGDUND")
}
}
