% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deployAppSaio}
\alias{deployAppSaio}
\title{Deploy metacheck to shinyapps.io
Wraps \link[rsconnect:deployApp]{rsconnect::deployApp}}
\usage{
deployAppSaio()
}
\description{
Deploy metacheck to shinyapps.io
Wraps \link[rsconnect:deployApp]{rsconnect::deployApp}
}
\note{
This will only work if metacheck and all its dependencies have been installed with \code{remotes::install_github()}
}
\seealso{
Other helpers: 
\code{\link{accumulate_pred_trans}()},
\code{\link{auth_cr}()},
\code{\link{auth_mailjet}()}
}
\concept{helpers}
