% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{.batch_query}
\alias{.batch_query}
\title{Batch query}
\usage{
.batch_query(urls, batch_size = 5, msg = "Batch Query", delay = 0.5)
}
\arguments{
\item{urls}{A vector of URLs}

\item{batch_size}{Size of each batch}

\item{msg}{Message to show in progress bar}

\item{delay}{Courtesy delay between batches (in seconds)}
}
\value{
a list of responses
}
\description{
Batch query
}
\keyword{internal}
