% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-grobid.R
\name{.grobid_to_bibr}
\alias{.grobid_to_bibr}
\title{Convert grobid to Bibr format}
\usage{
.grobid_to_bibr(xml_path, pb = NULL)
}
\arguments{
\item{xml_path}{path to a single XML file}

\item{pb}{a progress bar passed from \code{grobid_to_bibr()}}
}
\value{
a paper object
}
\description{
Convert grobid to Bibr format
}
\keyword{internal}
