% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-osf.R
\name{.osf_info}
\alias{.osf_info}
\title{Retrieve info from the OSF by ID}
\usage{
.osf_info(osf_ids, pb = NULL)
}
\arguments{
\item{osf_ids}{an vector of OSF IDs or URLs}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from the OSF by ID
}
\keyword{internal}
