% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobid.R
\name{.tei_xrefs}
\alias{.tei_xrefs}
\title{Get cross references from text table}
\usage{
.tei_xrefs(text_table)
}
\arguments{
\item{text_table}{The text table for the paper}
}
\value{
xrefs table
}
\description{
Get cross references from text table
}
\keyword{internal}
