% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-dryad.R
\name{dryad_info}
\alias{dryad_info}
\title{Retrieve info from Dryad by URL}
\usage{
dryad_info(dryad_url, id_col = 1, pb = NULL)
}
\arguments{
\item{dryad_url}{a Dryad dataset URL or DOI, or a table containing them
(e.g., as created by \code{\link[=dryad_links]{dryad_links()}})}

\item{id_col}{the index or name of the column that contains Dryad URLs, if
\code{dryad_url} is a table}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from Dryad by URL
}
\examples{
\dontrun{
  dryad_info("https://doi.org/10.5061/dryad.j1fd7")
}
}
