% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{email}
\alias{email}
\title{Set or get email}
\usage{
email(email = NULL)
}
\arguments{
\item{email}{if a string, sets the email}
}
\value{
the current option value (character)
}
\description{
Set or get email
}
\examples{
email()
}
