% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{emailReport}
\alias{emailReport}
\alias{emailReportUI}
\alias{emailReportServer}
\title{Email Report through a Shiny Module}
\usage{
emailReport()

emailReportUI(id, width = "100\%", ...)

emailReportServer(id, dois, email = blastula::prepare_test_message())
}
\arguments{
\item{id}{The id string to be namespaced (optional).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:reexports]{validateCssUnit()}}.}

\item{...}{
  Arguments passed on to \code{\link[shiny:actionButton]{shiny::actionButton}}
  \describe{
    \item{\code{inputId}}{The \code{input} slot that will be used to access the value.}
    \item{\code{label}}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}
    \item{\code{icon}}{An optional \code{\link[shiny:icon]{icon()}} to appear on the button.}
  }}
}
\description{
Email Report through a Shiny Module
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{emailReport}: Test app
}
\itemize{
\item \code{emailReportUI}: Module UI
}
\itemize{
\item \code{emailReportServer}: Module server
}}}

\seealso{
Other communicate: 
\code{\link{render_email}()},
\code{\link{runMetacheck}()},
\code{\link{smtp_send_metacheck}()}
}
\concept{communicate}
\keyword{internal}
