% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_funder.R
\name{funder_metrics}
\alias{funder_metrics}
\title{Funder metrics}
\usage{
funder_metrics(funder_info = NULL)
}
\arguments{
\item{funder_info}{tibble obtained with \code{\link[=cr_funder_df]{cr_funder_df()}} or \code{\link[=cr_compliance_overview]{cr_compliance_overview()}}}
}
\description{
Presents funders linked to publications in Crossref metadata
}
\details{
Publishers can share more than one funder link per record using Crossref metadata.
This function applies full counting, where the funder is counted onced per article.
Relative number is calculated by the number of checked articles.
Articles without funder information are tagged as "No funder links found".
}
\examples{
\dontrun{
## Some DOIs from OA articles
my_dois <- c("10.5194/wes-2019-70", "10.1038/s41598-020-57429-5",
  "10.3389/fmech.2019.00073", "10.1038/s41598-020-62245-y",
  "10.1109/JLT.2019.2961931")

# Workflow:
# First, obtain metadata from Crossref API
req <- get_cr_md(my_dois)

# Then, check article-level compliance

# Obtain funder metrics overview
funder_metrics(out$funder_info)
}
}
\keyword{internal}
