% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_cc.R
\name{get_compliant_cc}
\alias{get_compliant_cc}
\title{Extract records with compliant CC license metadata}
\usage{
get_compliant_cc(license_df)
}
\arguments{
\item{license_df}{data frame with licenses}
}
\description{
Extract records with compliant CC license metadata
}
\details{
In order to be compliant,
CC license has to apply to version of record and must be valid
without delay.
}
\keyword{internal}
