% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get_cr_md}
\alias{get_cr_md}
\title{Get Crossref metadata from API}
\usage{
get_cr_md(dois, .progress = ifelse(interactive(), "text", "none"))
}
\arguments{
\item{dois}{character vector with DOIs}

\item{.progress}{Show a \code{plyr}-style progress bar? Options are "none", "text",
"tk", "win", and "time".  See \code{\link[plyr]{create_progress_bar}} for details
of each. Only used when passing in multiple ids (e.g., multiple DOIs, DOI prefixes,
etc.), or when using the \code{cursor} param. When using the \code{cursor} param,
this argument only accept a boolean, either \code{TRUE} or \code{FALSE}; any
non-boolean is coerced to \code{FALSE}.}
}
\description{
Get Crossref metadata from API
}
\concept{ETL import}
