% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get_cr_md}
\alias{get_cr_md}
\title{Get Crossref metadata from API}
\usage{
get_cr_md(x)
}
\arguments{
\item{x}{A vector created by, or convertable to \code{\link[biblids:doi]{doi()}}.}
}
\description{
Get Crossref metadata from API
}
\seealso{
Other import: 
\code{\link{cr_works2}()},
\code{\link{pretest}}
}
\concept{import}
