% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_cc.R
\name{get_license_md}
\alias{get_license_md}
\title{Extract and normalize license info from Crossref metadata}
\usage{
get_license_md(cr)
}
\arguments{
\item{cr}{crossref metadata using \code{\link[=get_cr_md]{get_cr_md()}}}
}
\description{
Extract and normalize license info from Crossref metadata
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_tdm_df}()},
\code{\link{license_check}()},
\code{\link{vor_issue}()}
}
\concept{transform}
