% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_apc_fetch.R
\name{gh_file}
\alias{gh_file}
\title{Gets a file from a github repo, using the Data API blob endpoint}
\usage{
gh_file(
  url = NULL,
  ref = NULL,
  owner = NULL,
  repo = NULL,
  path = NULL,
  to_disk = TRUE,
  destfile = NULL,
  .token = NULL,
  .api_url = NULL,
  .method = "GET",
  .limit = NULL,
  .send_headers = NULL
)
}
\arguments{
\item{url}{the URL of the file to download via API, of the form
`github.com/:owner/:repo/blob/:path}

\item{owner, repo, path, ref}{alternate way to specify the file.  These will
override values in \code{url}}

\item{to_disk, destfile}{write file to disk (default=TRUE)?  If so, use the
name in \code{destfile}, or the original filename by default}

\item{.token}{Authentication token. Defaults to \code{GITHUB_PAT} or
\code{GITHUB_TOKEN} environment variables, in this order if any is set.
See \code{\link[gh:gh_token]{gh_token()}} if you need more flexibility, e.g. different tokens
for different GitHub Enterprise deployments.}

\item{.api_url}{Github API url (default: \url{https://api.github.com}). Used
if \code{endpoint} just contains a path. Defaults to \code{GITHUB_API_URL}
environment variable if set.}

\item{.method}{HTTP method to use if not explicitly supplied in the
\code{endpoint}.}

\item{.limit}{Number of records to return. This can be used
instead of manual pagination. By default it is \code{NULL},
which means that the defaults of the GitHub API are used.
You can set it to a number to request more (or less)
records, and also to \code{Inf} to request all records.
Note, that if you request many records, then multiple GitHub
API calls are used to get them, and this can take a potentially
long time.}

\item{.send_headers}{Named character vector of header field values
(except \code{Authorization}, which is handled via \code{.token}). This can be
used to override or augment the default \code{User-Agent} header:
\code{"https://github.com/r-lib/gh"}.}
}
\value{
Either the local path of the downloaded file (default), or a raw
vector
}
\description{
This avoids the 1MB limit of the content API and uses \link[gh:gh]{gh::gh} to deal with
authorization and such.  See https://developer.github.com/v3/git/blobs/
}
\author{
Noam Ross
}
