% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-grobid.R
\name{grobid_to_bibr}
\alias{grobid_to_bibr}
\title{Convert Grobid TEI XML file to bibr format}
\usage{
grobid_to_bibr(xml_file, save_path = ".", crossref_lookup = FALSE)
}
\arguments{
\item{xml_file}{the XML file}

\item{save_path}{directory or file path to save to; set to NULL to return a paper object}

\item{crossref_lookup}{whether to look up references in crossref}
}
\value{
a paper object
}
\description{
Convert Grobid TEI XML file to bibr format
}
