% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_ind.R
\name{ind_table_to_gt}
\alias{ind_table_to_gt}
\title{GT representation of compliance overview table}
\usage{
ind_table_to_gt(ind_table, .color = NULL)
}
\arguments{
\item{ind_table}{tibble compliance metrics table}

\item{.name}{indicator}

\item{.value}{number of articles per indicator}

\item{.prop}{percentage of articles per indicator}
}
\description{
GT representation of compliance overview table
}
\concept{visualize}
