% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_table.R
\name{info_table}
\alias{info_table}
\title{Get paper information in a table}
\usage{
info_table(
  paper,
  info = c("filename", "title", "keywords", "doi"),
  path = c("relative", "absolute")
)
}
\arguments{
\item{paper}{a paper object or a list of paper objects}

\item{info}{a vector of columns to return}

\item{path}{whether to return absolute or relative path for the filename}
}
\value{
a data frame with each paper id and info columns
}
\description{
Get paper information in a table
}
\examples{
info_table(psychsci[1:10])
}
