% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{lastlog}
\alias{lastlog}
\title{Get the last log}
\usage{
lastlog(i = 1, logpath = NULL)
}
\arguments{
\item{i}{the indices to return}

\item{logpath}{an optional file path to read the log from}
}
\value{
a list of the last log item, or a data frame of multiple items
}
\description{
Get the last log
}
\examples{
# set up 2 log items
logger("test", list(msg = "hi"))
logger("test", list(msg = "hi again"))

lastlog()
lastlog(2)
lastlog(1:2)
}
