% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_fetch.R
\name{license_normalise}
\alias{license_normalise}
\title{Normalise license info from Crossref}
\usage{
license_normalise(cr)
}
\arguments{
\item{cr}{tibble with non-compliant license metadata retrieved from the
Crossref API with \code{rcrossref:.cr_works()}}
}
\description{
Normalise license info from Crossref
}
