% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_fetch.R
\name{license_report}
\alias{license_report}
\title{Obtain records with non-compliant license information}
\usage{
license_report(cr)
}
\arguments{
\item{cr}{tibble with non-compliant license metadata retrieved from the Crossref API with \code{rcrossref:.cr_works()}}
}
\description{
In case license metadata do not comply, what are the reasons:
\itemize{
\item Did the publisher provide license metadata for the article?
\item Is the article provided under a CC license?
\item Did the CC license apply immediately after publication?
}
}
\seealso{
Other transform: 
\code{\link{cr_compliance_overview}()},
\code{\link{cr_funder_df}()},
\code{\link{cr_has_orcid}()},
\code{\link{cr_license_ind}()},
\code{\link{cr_md_ind}()},
\code{\link{cr_tdm_df}()},
\code{\link{cr_tdm_ind}()},
\code{\link{get_compliant_cc}()},
\code{\link{get_license_md}()},
\code{\link{license_check}()},
\code{\link{license_normalise}()},
\code{\link{license_val}()},
\code{\link{open_apc_get_}()},
\code{\link{open_apc_get}()},
\code{\link{vor_issue}()}
}
\concept{transform}
