% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_fetch.R
\name{license_val}
\alias{license_val}
\title{License checker}
\usage{
license_val(cr)
}
\arguments{
\item{cr}{tibble with Crossref metadata retrieved from
API with \code{rcrossref:.cr_works()}}
}
\description{
Retrieves records from Crossref metadata where
the version of record (vor), i.e. publisher version
is provided under a Creative Commons license without delay.
}
