% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm}
\alias{llm}
\title{Query an LLM}
\usage{
llm(
  text,
  query,
  text_col = "text",
  API_KEY = Sys.getenv("GROQ_API_KEY"),
  model = "llama3-70b-8192",
  maxTokens = 1024,
  temperature = 0,
  top_p = 1,
  proxy = NULL,
  include_query = FALSE
)
}
\arguments{
\item{text}{The text to send to the LLM (vector of strings, or data frame with the text in a column)}

\item{query}{The query to ask of the LLM}

\item{text_col}{The name of the text column if text is a data frame}

\item{API_KEY}{your API key for the LLM}

\item{model}{the LLM model name (see https://console.groq.com/docs/models)}

\item{maxTokens}{The maximum integer of completion tokens returned}

\item{temperature}{The amount of randomness in the response, valued between 0 inclusive and 2 exclusive}

\item{top_p}{Nucleus sampling threshold (between 0 and 1)}

\item{proxy}{NULL}

\item{include_query}{Whether to include the query string in the returned table}
}
\value{
a list of results
}
\description{
Ask a large language model (LLM) any question you want about the text from a search_text()
}
\details{
You may need to use your own API key. To avoid having to type it out, add it to the .Renviron file in the following format (you can use \code{usethis::edit_r_environ()} to access the .Renviron file)

CHATGPT_KEY="key_value_asdf"
}
\examples{
\donttest{
  text <- c("hello", "number", "ten", "12")
  query <- "Is this a number? Answer only 'yes' or 'no'"
  is_number <- llm(text, query)
}
}
