% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_use}
\alias{llm_use}
\title{Set or get metacheck LLM use}
\usage{
llm_use(llm_use = NULL)
}
\arguments{
\item{llm_use}{if logical, sets whether to use LLMs}
}
\value{
the current option value (logical)
}
\description{
Mainly for use in optional LLM workflows in modules, also checks if the GROQ API key is set and returns false if it isn't.
}
\examples{
if (llm_use()) {
  print("We can use LLMs")
} else {
  print("We will not use LLMs")
}
}
