% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logger}
\alias{logger}
\title{Log messages}
\usage{
logger(label = "", contents = list(), logpath = NULL)
}
\arguments{
\item{label}{a string with the context (e.g.,module name)}

\item{contents}{a named list of the log contents}

\item{logpath}{an optional file path to save the log in}
}
\value{
called for side effects of writing to log, returns logpath
}
\description{
Adds a logging message to the log. Keeps the log as a maximum of 1000 rows.
}
\examples{
logpath <- tempfile(fileext = ".log")
logger("test", list(x = 1), logpath)
lastlog()
}
