% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logpath}
\alias{logpath}
\title{Get log path}
\usage{
logpath()
}
\value{
the log file path
}
\description{
Checks the most recent log file created that day and creates a new one if missing.
}
\keyword{internal}
