% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{md_data_attachment}
\alias{md_data_attachment}
\title{Make Spreadsheet attachment}
\usage{
md_data_attachment(my_df = NULL, dois = NULL, session_id = NULL)
}
\arguments{
\item{my_df}{compliance data from \code{\link[=cr_compliance_overview]{cr_compliance_overview()}}}

\item{dois}{character, submitted dois}

\item{session_id}{link spreadsheet to R session}
}
\description{
Make Spreadsheet attachment
}
