% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{md_data_attachment}
\alias{md_data_attachment}
\title{Make Spreadsheet attachment}
\usage{
md_data_attachment(.md = NULL, session_id = NULL, dois = NULL)
}
\arguments{
\item{.md}{list, returned from \code{\link[=cr_compliance_overview]{cr_compliance_overview()}}}

\item{session_id}{link spreadsheet to R session}

\item{dois}{character, submitted dois}
}
\description{
Make Spreadsheet attachment
}
