% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacheck_package.R
\docType{package}
\name{metacheck-package}
\alias{metacheck}
\alias{metacheck-package}
\title{metacheck: Crossref Metadata Compliance}
\description{
Automatically check metadata compliance for hybrid open access (OA).
    Includes a webapp to send compliance reports by email.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://subugoe.github.io/metacheck}
  \item \url{https://github.com/subugoe/metacheck}
  \item Report bugs at \url{https://github.com/subugoe/metacheck/issues}
}

}
\author{
\strong{Maintainer}: Najko Jahn \email{najko.jahn@gmail.com} (\href{https://orcid.org/0000-0001-5105-1463}{ORCID})

Authors:
\itemize{
  \item Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})
  \item Henrieke Walter \email{henrieke.walter@sub.uni-goettingen.de} (\href{https://orcid.org/0000-0003-3313-7364}{ORCID})
}

Other contributors:
\itemize{
  \item Göttingen State and University Library [copyright holder, funder]
  \item German Research Foundation [sponsor]
  \item Crossref [data contributor]
}

}
\keyword{internal}
