% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{metacheck_app}
\alias{metacheck_app}
\title{Launch Demo App}
\usage{
metacheck_app(paper = NULL, quiet = FALSE, ...)
}
\arguments{
\item{paper}{optional paper or paperlist to load}

\item{quiet}{whether to show debugging messages in the console}

\item{...}{arguments to pass to shiny::runApp}
}
\value{
A paper object (invisibly)
}
\description{
Launch the demo app with Upload/Demo, full module selection, and text search.
}
\examples{
\dontrun{
metacheck_app()
}

}
