% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_overview.R
\name{metrics_overview}
\alias{metrics_overview}
\title{Metacheck metrics overview}
\usage{
metrics_overview(cr_overview = NULL)
}
\arguments{
\item{cr_overview}{tibble obtained with \code{\link[=cr_compliance_overview]{cr_compliance_overview()}}}
}
\description{
Presents the metacheck core compliance measures
}
\details{
The following metrics are presented (absolute and relative):
\itemize{
\item Creative commons license availability (\verb{CC License})
\item Compliant creative commons license availability (\verb{Compliant CC})
\item Support of links to full-texts for text and data mining (\verb{TDM Support})
\item Metadata records with links between publications and funder information (\verb{Funder info})
\item Metadata records with links between publications and ORCID  (\code{ORCID})
\item Metadata records with \href{https://i4oa.org/}{Open Abstracts}
\item Metadata records with \href{https://i4oc.org/}{Open Citations}
}
}
\examples{
\dontrun{
## Some DOIs from OA articles
my_dois <- c("10.5194/wes-2019-70", "10.1038/s41598-020-57429-5",
  "10.3389/fmech.2019.00073", "10.1038/s41598-020-62245-y",
  "10.1109/JLT.2019.2961931")

# Workflow:
# First, obtain metadata from Crossref API
req <- get_cr_md(my_dois)

# Then, check article-level compliance

# Finally, obtain compliance metrics overview
metrics_overview(out)
}
}
\keyword{internal}
