% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_run}
\alias{module_run}
\title{Run a module}
\usage{
module_run(paper, module, ...)
}
\arguments{
\item{paper}{a paper object or a list of paper objects}

\item{module}{the name of a module or path to a module to run on this object}

\item{...}{further arguments to the module (e.g., arguments for the \code{llm()} function like \code{seed}); these will override any arguments in the module}
}
\value{
a list of the returned table and report text
}
\description{
Run a module
}
\examples{
module_run(psychsci[[1]], "all_p_values")
}
