% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_apc_fetch.R
\name{open_apc_get}
\alias{open_apc_get}
\title{Get most current Open APC snapshot}
\usage{
open_apc_get(open_apc_cols = NULL)
}
\arguments{
\item{open_apc_cols}{character vector representing Open APC collections.
See \code{open_apc_collections()} for an overview.}
}
\value{
Tibble with article-level subset of Open APC dataset including
collection.
}
\description{
Open APC shares several csv files via GitHub, which contain data about
institutional spending for open access articles.
}
