% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_apc_fetch.R
\name{open_apc_get_}
\alias{open_apc_get_}
\title{Get single Open APC dataset using GitHub Data API}
\usage{
open_apc_get_(gh_link = NULL)
}
\arguments{
\item{gh_link}{URL to GitHub blob}
}
\description{
Get single Open APC dataset using GitHub Data API
}
