% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_file_download}
\alias{osf_file_download}
\title{Download all OSF Project Files}
\usage{
osf_file_download(
  osf_id,
  download_to = ".",
  max_file_size = 1,
  max_download_size = 10,
  max_folder_length = 40,
  ignore_folder_structure = FALSE
)
}
\arguments{
\item{osf_id}{an OSF ID or URL}

\item{download_to}{path to download to}

\item{max_file_size}{maximum file size to download (in MB) - set to NULL for no restrictions}

\item{max_download_size}{maximum total size to download (largest files will be omitted until total size is under the limit)}

\item{max_folder_length}{maximum folder name length}

\item{ignore_folder_structure}{if TRUE, download all files into a single folder}
}
\value{
folder structure
}
\description{
Download all OSF Project Files
}
\examples{
\donttest{
  osf_file_download("6nt4v")
}
}
