% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_get_all_pages}
\alias{osf_get_all_pages}
\title{Get All OSF API Query Pages}
\usage{
osf_get_all_pages(url, page_start = 1, page_end = Inf)
}
\arguments{
\item{url}{the OSF API URL}

\item{page_start}{the first page to get}

\item{page_end}{the last page to get (if Inf, then the final page)}
}
\value{
a table of the returned data
}
\description{
OSF API queries only return up to 10 items per page, so this helper functions checks for extra pages and returns all of them
}
\keyword{internal}
