% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-osf.R
\name{osf_links}
\alias{osf_links}
\title{Find OSF Links in Papers}
\usage{
osf_links(paper)
}
\arguments{
\item{paper}{a paper object or paperlist object}
}
\value{
a table with the OSF url in the first (href) column
}
\description{
Get all OSF links.
}
\examples{
osf_links(psychsci)
}
