% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper_write}
\alias{paper_write}
\title{Write paper}
\usage{
paper_write(paper, file_name = NULL, save_path = ".")
}
\arguments{
\item{paper}{a paper object}

\item{file_name}{the name of the file (if NULL, defaults to the paper_id)}

\item{save_path}{the directory to save the JSON file in}
}
\value{
the path to the JSON file
}
\description{
Save a paper as a JSON file.
}
\examples{
\dontrun{
paper <- demopaper()
paper$info$title <- "New title"
paper_write(paper, "new_paper")
}
}
