% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2grobid.R
\name{pdf2grobid}
\alias{pdf2grobid}
\title{Convert a PDF to Grobid XML}
\usage{
pdf2grobid(
  filename,
  save_path = ".",
  grobid_url = "https://kermitt2-grobid.hf.space",
  start = -1,
  end = -1,
  consolidate_citations = 0,
  consolidate_header = 0,
  consolidate_funders = 0,
  consolidateCitations = NULL,
  consolidateHeader = NULL,
  consolidateFunders = NULL
)
}
\arguments{
\item{filename}{path to the PDF, a vector of paths, or a directory name that contains PDFs}

\item{save_path}{directory or file path to save to; set to NULL to save to a temp file}

\item{grobid_url}{the URL to the grobid server}

\item{start}{the first page of the PDF to read (defaults to -1 to read all pages)}

\item{end}{the last page of the PDF to read (defaults to -1 to read all pages)}

\item{consolidate_citations}{whether to fix/enhance citations}

\item{consolidate_header}{whether to fix/enhance paper info}

\item{consolidate_funders}{whether to fix/enhance funder info}
}
\value{
XML object
}
\description{
This function uses a public grobid server maintained by Patrice Lopez. You can set up your own local grobid server following instructions from \url{https://grobid.readthedocs.io/} and set the argument \code{grobid_url} to its path (probably \url{http://localhost:8070})
}
\details{
Consolidation of citations, headers, and funders looks up these items in CrossRef or another database to fix or enhance information (see \url{https://grobid.readthedocs.io/en/latest/Consolidation/}). This can slow down conversion. Consolidating headers is only useful for published papers, and can be set to 0 for work in prep.
}
