% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-psycharchives.R
\name{psycharchives_info}
\alias{psycharchives_info}
\title{Retrieve info from PsychArchives by URL}
\usage{
psycharchives_info(pa_url, id_col = 1, pb = NULL)
}
\arguments{
\item{pa_url}{a PsychArchives URL (item page, handle, or hdl.handle.net link),
or a table containing them (e.g., as created by \code{psycharchives_links()})}

\item{id_col}{the index or name of the column that contains PsychArchives URLs,
if \code{pa_url} is a table}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Retrieve info from PsychArchives by URL
}
\examples{
\dontrun{
psycharchives_info("https://hdl.handle.net/20.500.12034/17526")
}
}
