% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\alias{read_grobid}
\alias{read_cermine}
\alias{read_text}
\title{Get paper from XML or text file}
\usage{
read(filename)

read_grobid(filename)

read_cermine(filename)

read_text(filename)
}
\arguments{
\item{filename}{the path to the file, a vector of file paths, or the path to a directory}
}
\value{
A paper object with class scivrs_paper, or a list of paper objects
}
\description{
This should work with XML files in TEI (grobid), JATS APA-DTD, NLM-DTD and cermine formats, plus basic parsing of plain text files.
}
\examples{
# paper <- read(filename)
}
