% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibr.R
\name{read}
\alias{read}
\title{Read in grobis XML or bibr ZIP}
\usage{
read(file_path)
}
\arguments{
\item{file_path}{path to a directory containing XML and/or zip files, or a vector of paths to XML or zip files}
}
\value{
a paper or paperlist
}
\description{
Read in grobis XML or bibr ZIP
}
