% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibr.R
\name{read_bibr}
\alias{read_bibr}
\title{Read bibr JSON file}
\usage{
read_bibr(file_path)
}
\arguments{
\item{file_path}{path to the JSON file (or legacy ZIP file)}
}
\value{
a paper object
}
\description{
Read bibr JSON file
}
\keyword{internal}
