% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{ref_table}
\alias{ref_table}
\title{Reference and DOI table}
\usage{
ref_table(paper)
}
\arguments{
\item{paper}{a paper or paperlist}
}
\value{
a merged table
}
\description{
Return a table with fixed DOIs and reference text from a paper object or concatenate tables across a list of paper objects.
}
\examples{
biblio <- ref_table(psychsci[[1]])
}
