% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{render_email}
\alias{render_email}
\title{Render email}
\usage{
render_email(dois, session_id = NULL)
}
\arguments{
\item{dois}{Character vector of DOIs}

\item{session_id}{Character vector to identify current shiny session}
}
\description{
Render email
}
\seealso{
Other communicate: 
\code{\link{emailReport}()},
\code{\link{smtp_send_metacheck}()}
}
\concept{communicate}
