% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{render_email}
\alias{render_email}
\alias{render_and_send}
\alias{render_and_send_async}
\title{Render email}
\usage{
render_email(dois, lang = "en", session_id = NULL)

render_and_send(dois, to, lang)

render_and_send_async(dois, to, lang)
}
\arguments{
\item{dois}{Character vector of DOIs}

\item{lang}{Character scalar giving the anguage of the report.
\emph{Not} reactive as in \code{\link[=mcControlsServer]{mcControlsServer()}}.}

\item{session_id}{Character vector to identify current shiny session}

\item{to}{A vector of email addresses serving as primary recipients for the
message. For secondary recipients, use the \code{cc} and \code{bcc} arguments. A
named character vector can be used to specify the recipient names along
with the their email address (e.g., \code{c("Jane Doe" = "jane_doe@example.com")}).}
}
\description{
Render email
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{render_and_send}: Render and send
}
\itemize{
\item \code{render_and_send_async}: Render and send asynchronously
}}}

\seealso{
Other communicate: 
\code{\link{emailReport}()},
\code{\link{mcApp}()},
\code{\link{mcControls}},
\code{\link{report}},
\code{\link{runMetacheck}()},
\code{\link{smtp_send_metacheck}()}
}
\concept{communicate}
