% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-4tu.R
\name{researchdata4tu_info}
\alias{researchdata4tu_info}
\title{Retrieve info from 4TU.ResearchData by URL}
\usage{
researchdata4tu_info(researchdata4tu_url, id_col = 1, pb = NULL)
}
\arguments{
\item{researchdata4tu_url}{a 4TU.ResearchData URL or DOI, or a table
containing them (e.g., as created by \code{\link[=researchdata4tu_links]{researchdata4tu_links()}})}

\item{id_col}{the index or name of the column that contains
4TU.ResearchData URLs, if \code{researchdata4tu_url} is a table}

\item{pb}{a progress bar passed from another function}
}
\value{
a data frame of information
}
\description{
Thin wrapper around \code{\link[=figshare_info]{figshare_info()}} with \code{host = "data.4tu.nl"} -- see
the note at the top of this file for why 4TU.ResearchData's Djehuty
platform can reuse the Figshare-family implementation directly.
}
\examples{
\dontrun{
  researchdata4tu_info("https://doi.org/10.4121/16766929.v1")
}
}
