% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_tdm.R
\name{tdm_metrics}
\alias{tdm_metrics}
\title{TDM MIME-Type metrics}
\usage{
tdm_metrics(tdm = NULL)
}
\arguments{
\item{tdm}{tibble obtained with \code{\link[=cr_tdm_df]{cr_tdm_df()}} or \code{\link[=cr_compliance_overview]{cr_compliance_overview()}}}
}
\description{
Presents MIME-Types for full-text versions of version-of-records.
}
\details{
Publishers can share more than one TDM-link per full-text using
Crossref metadata. This function applies full counting, where
full-texts were assigned to more than one MIME type were counted once
for each. Relative number is calculated by the number of checked articles.
Articles without compliant TDM information are tagged as
"No TDM links for version-of-records".
}
\examples{
# First, obtain metadata from Crossref API
req <- get_cr_md(doi_examples$good)

# Then, check article-level compliance
out <- cr_compliance_overview(req)

# Finally, obtain TDM metrics overview
tdm_metrics(out$tdm)
}
\keyword{internal}
