% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{test_paper}
\alias{test_paper}
\title{Test paper}
\usage{
test_paper(text, id = "test")
}
\arguments{
\item{text}{a vector of text to add}

\item{id}{the ID for the paper}
}
\value{
a paper object
}
\description{
Create a paper object with the specified text and id (mainly for testing/demos).
}
\examples{
# to test a paper with a specific URL
p <- test_paper("https://osf.io/abcde")
}
