% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{text_bib}
\alias{text_bib}
\title{Get references/citations from plain text}
\usage{
text_bib(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
an list of two tables: references and citations
}
\description{
Get references/citations from plain text
}
\keyword{internal}
