% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate a module}
\usage{
validate(module, sample, expected = NULL, path = NULL)
}
\arguments{
\item{module}{The name of a module or path to a module}

\item{sample}{A data frame or path to a tabular data file (csv or excel) for the sample to validate against, with required column id, and optional columns table, report, traffic_light}

\item{expected}{A data frame or the path to a file that contains the expected results table values in tabular form; if NULL, the table column of the sample table will be used or table checks omitted}

\item{path}{The base directory for relative paths in the sample table; if NULL and the sample is a file path, it will use the base directory of the file}
}
\value{
A validation list object
}
\description{
Validate a module
}
