% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate}
\usage{
validate(gt, module, compare = "table")
}
\arguments{
\item{gt}{a data frame or vector of text}

\item{module}{the module}

\item{compare}{name of the module output table for comparison}
}
\value{
something
}
\description{
Validate
}
\examples{
validate("p < .05", "stat_p_exact")
}
