% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_md_cc.R
\name{vor_issue}
\alias{vor_issue}
\title{Extract records with CC license not applied to version of records}
\usage{
vor_issue(license_df, compliant_dois)
}
\arguments{
\item{license_df}{normalized license metadata from \code{\link[=license_check]{license_check()}}}

\item{compliant_dois}{DOIs representing records with valid CC info}
}
\description{
Extract records with CC license not applied to version of records
}
\details{
In order to be compliant,
CC license has to apply to version of record and must be valid
without delay.
}
