% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{word_bib}
\alias{word_bib}
\title{Get bibliography from Word .docx}
\usage{
word_bib(xml)
}
\arguments{
\item{xml}{The XML}
}
\value{
bib table
}
\description{
Get bibliography from Word .docx
}
\keyword{internal}
