% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-zenodo.R
\name{zenodo_links}
\alias{zenodo_links}
\title{Find Zenodo Links in Papers}
\usage{
zenodo_links(paper)
}
\arguments{
\item{paper}{a paper object or paperlist object}
}
\value{
a table with the Zenodo url in the first (text) column
}
\description{
Find Zenodo Links in Papers
}
\examples{
zenodo_links(psychsci)
}
